% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_labels.R
\name{apply_labels}
\alias{.apply_labels}
\alias{.apply_val_labs}
\alias{.apply_var_labs}
\alias{apply_labels}
\alias{apply_val_labs}
\alias{apply_var_labs}
\title{Set variable labels/value labels on variables in the data.frame}
\usage{
apply_labels(data, ...)

apply_var_labs(data, ...)

apply_val_labs(data, ...)

.apply_labels(...)

.apply_val_labs(...)

.apply_var_labs(...)
}
\arguments{
\item{data}{data.frame/list}

\item{...}{named arguments. Name of argument is a variable name in 
\code{data}. Arguments values is variable label/value labels. For 
\code{apply_labels} characters are considered as variable labels and named 
vectors are considered as value labels.}
}
\value{
\code{data} with applied labels
}
\description{
Set variable labels/value labels on variables in the data.frame
}
\examples{
1
}

