% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{as.labelled}
\alias{as.labelled}
\alias{is.labelled}
\title{Recode vector into numeric vector with value labels}
\usage{
as.labelled(x, label = NULL)

is.labelled(x)
}
\arguments{
\item{x}{numeric vector/character vector/factor}

\item{label}{optional variable label}
}
\value{
numeric vector with value labels
}
\description{
Recode vector into numeric vector with value labels
}
\examples{
character_vector = c("one", "two",  "two", "three")
as.labelled(character_vector, label = "Numbers")


}
