% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_hux.R
\name{as_huxtable}
\alias{as_huxtable}
\alias{as_hux}
\alias{as_huxtable.etable}
\title{Convert table to huxtable}
\usage{
as_huxtable(x, ...)

as_hux(x, ...)

\method{as_huxtable}{etable}(x, ...)
}
\arguments{
\item{x}{etable. Table to convert to a huxtable.}
}
\description{
Convert table to huxtable
}
