% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{banner}
\alias{banner}
\title{Title}
\usage{
banner(..., weight = NULL, subset = NULL)
}
\arguments{
\item{...}{vectors/data.frames/lists. Variables which will be used as banner
of the table.}

\item{weight}{optional numeric vector - weight for tables}

\item{subset}{optional logical vector specifying a subset of observations to
be used for the table.}
}
\value{
list of variables
}
\description{
Title
}
\examples{
1 == 1
}

