% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/category.R
\name{category}
\alias{category}
\title{Convert dichotomy matrix to matrix with category encoding.}
\usage{
category(x, counted_value = 1, compress = TRUE)
}
\arguments{
\item{x}{Dichotomy matrix (usually with 0,1 coding).}

\item{counted_value}{Vector. Values that should be considered as indicator
of category presence. By default it is 1.}

\item{compress}{Logical. Should we drop columns with all NA?}
}
\value{
Matrix with numeric values that correspond to column numbers of counted values.
Column names of x are added as value labels.
}
\description{
Convert dichotomy matrix to matrix with category encoding.
}
\examples{
set.seed(123)

# Let's imagine it's matrix of consumed products
dichotomy_matrix = matrix(sample(0:1,40,replace = TRUE,prob=c(.6,.4)),nrow=10)
colnames(dichotomy_matrix) = c("Milk","Sugar","Tea","Coffee")
category(dichotomy_matrix,compress=FALSE) # uncompressed version
category_matrix=category(dichotomy_matrix)

 # should be TRUE
identical(val_lab(category_matrix),c(Milk = 1L,Sugar = 2L,Tea = 3L,Coffee = 4L))
all(dichotomy(category_matrix,use_na = FALSE)==dichotomy_matrix)
}
\seealso{
\code{\link{dichotomy}} for reverse conversion.
}

