% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_internal.R
\name{check_internal}
\alias{check_internal}
\alias{print.check}
\alias{summary.check}
\title{Internal function for checking multiple/single-response questions for valid values}
\usage{
check_internal(dfs, values = NULL, exclusive = NULL, mult = FALSE,
  no_dup = mult, cond = NULL, subset = NULL)

\method{print}{check}(x, error_num = 20, ...)

\method{summary}{check}(object, skip_details = FALSE)
}
\arguments{
\item{dfs}{Vector/data.frame/matrix that should be checked.}

\item{values}{Valid values. All other values will be considered incorrect.}

\item{exclusive}{Numeric/character values. These values should be exclusive.
All other values should be NA if any of exclusive values exists in row.}

\item{mult}{Logical. Should we check dfs as multiple response question?}

\item{no_dup}{Logical. Should we check for absence of duplicated values in each row?}

\item{cond}{Logical vector. TRUE indicated rows in dfs that should contain valid
values. In other rows all dfs values should be NA. It used for questions that
 were asked by condition on answer on previous questions.}

\item{subset}{Logical vector. TRUE indicated rows in dfs that should be checked.
Other rows will be ignored.}

\item{x}{Check object for printing.}

\item{error_num}{Numeric. How many errors should be printed?}

\item{object}{Check object for summary.}

\item{skip_details}{Logical. For more terse summary.}
}
\value{
\code{check_internal} return object of class 'check'. It is data.frame that contains
check result for each row of dfs and description for each error if any of them
 exists.
\code{print.check} invisibly returns its argument x.
\code{summary.check} returns list with summary check.
}
\description{
This is low-level function intended for programming. For interactive usage
see functions \code{\link{mult}} and \code{\link{sngl}}.
}
\details{
'mult=TRUE' for multiple response questions means that it is allowed to have NA
 values in row. It is only necessary for multiple response questions that each
 row will have at least one non-NA values.
 This function checks multiple response questions only with categorical coding.
 For checking multiple response questions with dichotomous coding see \code{\link{dmult}}.
 If 'mult=FALSE' all values should be non-NA. However one can put NA in 'values' argument.
 Then NA will be considered valid.
 By default if 'mult=TRUE' then no_dup also is TRUE.
 If 'values' is missing than all values considered valid except NA.
 'check' report only for first error in row. If there are other errors for
 this case they will be reported only after correction of first error.
}
\examples{
library(dplyr)
data(ProductTestRaw)

## Example 1 ##

# 4 errors: 2 missing, 2 invalid codes
check_internal(ProductTestRaw$s2b,values=2:3)

## Example 2 ##

data(codeframe)
valid_a1 = make_labels(codeframe$likes)

# Exclusive values
# 1 Liked everything
# 2 Disliked everything
# 99 Hard to say

# 5 errors: 1 missing value, 1 invalid code, 1 code duplication,
# 2 non-exclusive values
check_internal(select(ProductTestRaw,a1_1:a1_6),values=valid_a1,
     mult = TRUE, exclusive=c(1,2,99))

## Example 3 ##

valid_a4 = make_labels(codeframe$dislikes_in_appearance)
# question a4 was asked only if codes 1-4 marked in a3
# 3 errors: 1 missing value, 1 invalid code, 1 code in case of a3 in 5-7.
check_internal(select(ProductTestRaw,a4_1:a4_6),
     values=valid_a4,mult = TRUE, exclusive=99,
      cond = ProductTestRaw$a3 \%in\%  1:4)
}
\seealso{
\code{\link{mult}}, \code{\link{mult_}}, \code{\link{sngl}},
\code{\link{sngl_}}, \code{\link{dmult}}
}

