% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tables.R
\name{col_vars_}
\alias{col_vars_}
\alias{cell_}
\alias{row_vars_}
\alias{stat}
\alias{stat_}
\alias{weight_}
\alias{subgroup_}
\title{Functions for constracting custom tables}
\usage{
col_vars_(...)

cell_(..., cell_name = NULL)

row_vars_(...)

stat(..., margins = c("rows", "columns", "total"))

stat_(..., margins = c("rows", "columns", "total"))

weight_(...)

subgroup_(...)
}
\arguments{
\item{...}{jjjkj}
}
\value{
jkjk
}
\description{
Functions for constracting custom tables
}
\examples{
1
}
