% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_tests.R
\name{compare_proportions}
\alias{compare_proportions}
\alias{compare_means}
\title{Calculate significance (p-values) of differences between proportions/means}
\usage{
compare_proportions(prop1, prop2, base1, base2, common_base = 0)

compare_means(mean1, mean2, sd1, sd2, base1, base2, common_base = 0,
  var_equal = FALSE)
}
\arguments{
\item{prop1}{fsdfdf}

\item{prop2}{sfs}

\item{base1}{sfsf}

\item{base2}{sfsf}

\item{common_base}{sfsf}

\item{mean1}{fsdfdf}

\item{mean2}{sfs}

\item{sd1}{fsdfdf}

\item{sd2}{sfs}

\item{var_equal}{sfsf}
}
\value{
numeric vector with p-values
}
\description{
Calculate significance (p-values) of differences between proportions/means
}
\examples{
1
}
