% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_experimental.R
\name{compute}
\alias{..add_val_lab}
\alias{..compute}
\alias{..do_if}
\alias{..if_val}
\alias{..modify}
\alias{..modify_if}
\alias{..recode}
\alias{..set_val_lab}
\alias{..set_var_lab}
\alias{.cro}
\alias{.cro_cpct}
\alias{.cro_rpct}
\alias{.cro_tpct}
\alias{.fre}
\alias{.val_lab}
\alias{.var_lab}
\alias{compute}
\title{Experimental functions for working with default dataset}
\usage{
..modify(expr)

..modify_if(cond, expr)

..do_if(cond, expr)

..compute(expr)

.val_lab(...)

.var_lab(...)

..set_var_lab(x, ...)

..set_val_lab(x, ...)

..add_val_lab(x, ...)

..if_val(x, ...)

..recode(x, ...)

.fre(...)

.cro(...)

.cro_cpct(...)

.cro_rpct(...)

.cro_tpct(...)
}
\arguments{
\item{expr}{dfsdf}

\item{cond}{sdfsdf}

\item{...}{dsfsdf}

\item{x}{dsfs}
}
\description{
Experimental functions for working with default dataset
}
\examples{
a = 1
}

