% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cro.R
\name{cro}
\alias{cro}
\alias{cro_cases}
\alias{cro_cpct}
\alias{cro_rpct}
\alias{cro_tpct}
\alias{cro_cpct_responses}
\title{Cross tabulation with support of labels, weights and multiple response variables.}
\usage{
cro(cell_vars, col_vars = total(), row_vars = total(label = ""),
  weight = NULL, subgroup = NULL, total_label = NULL,
  total_statistic = "u_cases", total_row_position = c("below", "above",
  "none"))

cro_cases(cell_vars, col_vars = total(), row_vars = total(label = ""),
  weight = NULL, subgroup = NULL, total_label = NULL,
  total_statistic = "u_cases", total_row_position = c("below", "above",
  "none"))

cro_cpct(cell_vars, col_vars = total(), row_vars = total(label = ""),
  weight = NULL, subgroup = NULL, total_label = NULL,
  total_statistic = "u_cases", total_row_position = c("below", "above",
  "none"))

cro_rpct(cell_vars, col_vars = total(), row_vars = total(label = ""),
  weight = NULL, subgroup = NULL, total_label = NULL,
  total_statistic = "u_cases", total_row_position = c("below", "above",
  "none"))

cro_tpct(cell_vars, col_vars = total(), row_vars = total(label = ""),
  weight = NULL, subgroup = NULL, total_label = NULL,
  total_statistic = "u_cases", total_row_position = c("below", "above",
  "none"))

cro_cpct_responses(cell_vars, col_vars = total(), row_vars = total(label =
  ""), weight = NULL, subgroup = NULL, total_label = NULL,
  total_statistic = "u_responses", total_row_position = c("below", "above",
  "none"))
}
\arguments{
\item{cell_vars}{vector/data.frame/list.}

\item{col_vars}{vector/data.frame/list. By now multiple-response predictor is not supported.}

\item{row_vars}{vector/data.frame/list. By now multiple-response predictor is not supported.}

\item{weight}{numeric vector. Optional cases weights. Cases with NA's,
negative and zero weights are removed before calculations.}

\item{subgroup}{logical vector.}

\item{total_label}{By default "#Total". You can provide several names - each name for
each total statistics.}

\item{total_statistic}{By default it is "u_cases" (unweighted cases). 
Possible values are "u_cases", "u_responses", "u_cpct", "u_rpct", "u_tpct",
"w_cases", "w_responses", "w_cpct", "w_rpct", "w_tpct". "u_" means
unweighted statictics and "w_" means weighted statistics.}

\item{total_row_position}{Position ot total row in the resulting table. Can
be one of "below", "above", "none".}
}
\value{
object of class 'etable'. Basically it's a data.frame but class
  is needed for custom methods.
}
\description{
\itemize{
\item{\code{cro}}{ build a contingency table of the counts.}
\item{\code{cro_cpct}, \code{cro_cpct_responses}}{ build a contingency table 
of the column percents. These functions give different results only for 
multiple response variables. For \code{cro_cpct} base of percents is number 
of valid cases. Case is considered as valid if it has at least one non-NA 
value. So for multiple response variables sum of percents may be greater than
100. For \code{cro_cpct_responses} base of percents is number of valid 
responses. Multiple response variables can have several responses for single 
case. Sum of percents of \code{cro_cpct_responses} always equals to 100\%.}
\item{\code{cro_rpct}}{ build a contingency table of the row percents. Base
for percents is number of valid cases.}
\item{\code{cro_tpct}}{ build a contingency table of the table percents. Base
for percents is number of valid cases.}
}
To provide multiple-response variables as arguments use \link{mrset} for
multiples with category encoding and \link{mdset} for multiples with
dichotomy (dummy) encoding.
}
\examples{
data(mtcars)
mtcars = apply_labels(mtcars,
                      mpg = "Miles/(US) gallon|Mean",
                      cyl = "Number of cylinders",
                      disp = "Displacement (cu.in.)|Mean",
                      hp = "Gross horsepower|Mean",
                      drat = "Rear axle ratio",
                      wt = "Weight (lb/1000)",
                      qsec = "1/4 mile time|Mean",
                      vs = "Engine",
                      vs = c("V-engine" = 0,
                             "Straight engine" = 1),
                      am = "Transmission",
                      am = c("Automatic" = 0,
                             "Manual"=1),
                      gear = "Number of forward gears",
                      carb = "Number of carburetors"
)

calculate(mtcars, cro(am, vs))

# multiple banners
calculate(mtcars, cro(cyl, list(total(), vs, am, carb)))

# column percent
calculate(mtcars, cro_cpct(cyl, list(total(), vs, am, carb)))

# nested banner
calculate(mtcars, cro(cyl, list(total(), vs \%nest\% am, carb)))

# stacked variables
calculate(mtcars, cro(list(cyl, carb), list(total(), vs \%nest\% am)))

# nested variables
calculate(mtcars, cro(am \%nest\% cyl, list(total(), vs, carb)))

# multiple-choice variable
# brands - multiple response question
# Which brands do you use during last three months? 
set.seed(123)
brands = data.frame(t(replicate(20,sample(c(1:5,NA),4,replace = FALSE))))
# score - evaluation of tested product
score = sample(-1:1,20,replace = TRUE)
var_lab(brands) = "Used brands"
val_lab(brands) = make_labels("
                              1 Brand A
                              2 Brand B
                              3 Brand C
                              4 Brand D
                              5 Brand E
                              ")

var_lab(score) = "Evaluation of tested brand"
val_lab(score) = make_labels("
                             -1 Dislike it
                             0 So-so
                             1 Like it    
                             ")

cro(mrset(brands), list(total(), score))
cro_cpct(mrset(brands), list(total(), score))

}
\seealso{
\link{tables}, \link{fre}, \link{cro_fun}.
}
