% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_datatable.R
\name{datatable}
\alias{datatable}
\alias{datatable.simple_table}
\title{Title}
\usage{
datatable(data, ...)

\method{datatable}{simple_table}(data, repeat_row_labels = FALSE,
  show_row_numbers = FALSE, digits = NULL, ...)
}
\arguments{
\item{data}{a data object (either a matrix or a data frame)}

\item{...}{dfd}

\item{repeat_row_labels}{logical}

\item{show_row_numbers}{logical}

\item{digits}{integer}
}
\value{
fdf
}
\description{
Title
}
\examples{
a = 1
}
