% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{\%in_row\%}
\alias{\%in_col\%}
\alias{ml_left}
\alias{ml_right}
\alias{ml_autonum}
\alias{vars_range}
\alias{vars_range_list}
\alias{vars_pattern}
\alias{vars_pattern_list}
\alias{.with}
\alias{.set_var_lab}
\alias{.set_val_lab}
\alias{.add_val_lab}
\alias{category}
\alias{category_df}
\title{Deprecated functions (don't use them)}
\usage{
criterion \%in_row\% x

criterion \%in_col\% x

ml_left(text)

ml_right(text)

ml_autonum(text)

vars_range(start, end)

vars_range_list(start, end)

vars_pattern(pattern)

vars_pattern_list(pattern)

.with(expr, ...)

.set_var_lab(x, ...)

.set_val_lab(x, ...)

.add_val_lab(x, ...)

category(x, prefix = NULL, use_var_lab = TRUE, counted_value = 1,
  compress = FALSE)

category_df(x, prefix = NULL, use_var_lab = TRUE, counted_value = 1,
  compress = FALSE)
}
\arguments{
\item{criterion}{deprecated}

\item{x}{deprecated}

\item{text}{deprecated}

\item{start}{deprecated}

\item{end}{deprecated}

\item{pattern}{deprecated}

\item{expr}{deprecated}

\item{...}{deprecated}
}
\description{
Deprecated functions (don't use them)
}
