% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_empty.R
\name{drop_empty_rows}
\alias{drop_empty_columns}
\alias{drop_empty_rows}
\title{Title}
\usage{
drop_empty_rows(x, excluded_columns = 1, excluded_rows = grep("^#", x[[1]],
  perl = TRUE))

drop_empty_columns(x, excluded_columns = 1, excluded_rows = grep("^#",
  x[[1]], perl = TRUE))
}
\arguments{
\item{x}{fsdfsfd}

\item{excluded_columns}{dsfsdf}

\item{excluded_rows}{sdfsdf}
}
\value{
dfsdf
}
\description{
Title
}
\examples{
a = 1
}

