% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtfrm.R
\name{dtfrm}
\alias{as.dtfrm}
\alias{dtfrm}
\alias{lst}
\title{Make data.frame without conversion to factors and without fixing names}
\usage{
dtfrm(...)

as.dtfrm(x, ...)

lst(...)
}
\arguments{
\item{...}{objects, possibly named}

\item{x}{object to be coerced}
}
\value{
data.frame/list
}
\description{
\code{dtfrm} and \code{as.dtfrm} are shortcuts to \code{data.frame} and 
\code{as.data.frame} with stringsAsFactors = FALSE, check.names = FALSE.
\code{lst} creates list with names.
}
\examples{

# see the difference
df1 = data.frame(a = letters[1:3], "This is my long name" = 1:3)
df2 = dtfrm(a = letters[1:3], "This is my long name" = 1:3)

str(df1)
str(df2)

# lst
a = 1:3
b = 3:1

list1 = list(a, b)
list2 = lst(a, b)

str(list1)
str(list2)

}
\seealso{
\link[base]{data.frame}, \link[base]{as.data.frame},  \link[base]{list}
}

