% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/error_if.R
\name{error_if}
\alias{error_if}
\title{Function for checking arbitrary error conditions.}
\usage{
error_if(.data, expr, show = NULL)
}
\arguments{
\item{.data}{Dataset that will be checked.}

\item{expr}{Expression that returns logical vector. TRUE is incorrect value (error),
all other values are considered as valid cases.}

\item{show}{Additional variables (such as 'id') that should be shown with report about
errors.}
}
\value{
original data.frame .data with attribute with
  results of checking.
}
\description{
This function is designed for working with piping operator \code{\link[magrittr]{\%>\%}}
 from 'magrittr' package.
}
\details{
Supposed usage is: \code{checked_data.frame  \%>\% error_if(expression)}.
 It is possible to use this function with conditional operators \code{\link{check_if}}
 and \code{\link{check_subset}}. Both conditional operators will have the same effect:
 expression will be checked only for TRUE values in conditions.
}
\examples{
library(dplyr)
data(ProductTestRaw)

# Age category and age should be consistent. 4 errors
ProductTestRaw  \%>\% error_if((s2a \%in\% 18:26) & !(s2b \%in\% 2))  \%>\% report
# No errors
ProductTestRaw  \%>\% error_if((s2a \%in\% 27:35) & !(s2b \%in\% 3))  \%>\% report

# Show 'id' and 'cell'
ProductTestRaw  \%>\% error_if((s2a \%in\% 18:26) & !(s2b \%in\% 2),show=c(id,cell))  \%>\% report

# Conditional checking. There  are no errors in cell 2.

ProductTestRaw  \%>\% check_if(cell==2)  \%>\% error_if((s2a \%in\% 18:26) & !(s2b \%in\% 2))  \%>\% report

# Usage in programming (e. g. in cycle 'for')
# Similar to Example 4 for sngl function.
checked_vars = c("a3","a22","b3","b23")

# there is one error in a22

for (each_var in checked_vars){
     cat("Question:",each_var,"\\n")
     ProductTestRaw  \%>\%
         error_if(!(ProductTestRaw[,each_var] \%in\% 1:7),show=c(id,cell))  \%>\%
         report
}
}
\seealso{
\code{\link{check_if}}, \code{\link{check_subset}},
\code{\link[magrittr]{\%>\%}}, \code{\link{move}}
}

