% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factor.R
\name{factor}
\alias{as.factor}
\alias{as.ordered}
\alias{factor}
\alias{ordered}
\title{Factors}
\usage{
factor(x = character(), levels, labels = levels, exclude = NA,
  ordered = is.ordered(x), nmax = NA)

as.factor(x)

ordered(x, ...)

as.ordered(x)
}
\arguments{
\item{x}{a vector of data, usually taking a small number of distinct values.}

\item{levels}{an optional vector of the values (as character strings) that x might have taken. The default is the unique set of values taken by as.character(x), sorted into increasing order of x. Note that this set can be specified as smaller than sort(unique(x)).}

\item{labels}{either an optional character vector of labels for the levels (in the same order as levels after removing those in exclude), or a character string of length 1.}

\item{exclude}{a vector of values to be excluded when forming the set of levels. This should be of the same type as x, and will be coerced if necessary.}

\item{ordered}{logical flag to determine if the levels should be regarded as ordered (in the order given).}

\item{nmax}{an upper bound on the number of levels.}

\item{...}{(in ordered(.)): any of the above, apart from ordered itself.}
}
\value{
\code{factor}, \code{as.factor}, \code{ordered}, \code{as.ordered}
return an object of class factor. For details see base \code{factor} documentation.
}
\description{
These functions override the same functions provided in base to make them
generic so that methods for variables with labels can be
provided. The default methods call the base versions.
}
\details{
For variables with labels arguments \code{levels}, \code{labels}, \code{exclude},
 \code{nmax} are ignored. Factor levels are constructed as values from values labels + variable values
 without labels (so there is no information lost). Labels are constructed as "Variable_label|Value label".
}
\examples{
data(mtcars)

var_lab(mtcars$am) = "Transmission"
val_lab(mtcars$am) = c(automatic = 0, manual=1)

\dontrun{
plot(factor(mtcars$am))
}

summary(lm(mpg ~ am, data = mtcars)) # no labels
summary(lm(mpg ~ factor(am), data = mtcars)) # with labels
summary(lm(mpg ~ factor(unvr(am)), data = mtcars)) # without variable label
}
\seealso{
Materials for base functions: \code{\link[base]{factor}}, \code{\link[base]{as.factor}},
 \code{\link[base]{ordered}}, \code{\link[base]{as.ordered}}
}

