% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{htmlTable.simple_table}
\alias{htmlTable.simple_table}
\alias{knit_print.simple_table}
\alias{knit_print.etable}
\title{Outputting HTML tables}
\usage{
\method{htmlTable}{simple_table}(x, digits = getOption("expss.digits"), ...)

knit_print.simple_table(x, digits = NULL, ...)

knit_print.etable(x, digits = NULL, ...)
}
\arguments{
\item{x}{a data object (result of \link{fre}/\link{cro} and etc)}

\item{digits}{integer If it is not NULL than all numeric columns will be
rounded to specified number of digits.}

\item{...}{further parameters for \link[htmlTable]{htmlTable}}
}
\value{
Returns a string of class htmlTable
}
\description{
This is method for rendering results of \link{fre}/\link{cro} in Shiny.
For detailed description of function and its arguments see \link[DT]{datatable}.
}
\examples{
a = 1
}
