% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelr.R
\docType{package}
\name{labelr}
\alias{labelr}
\alias{labelr-package}
\title{excel.link: convenient data exchange with Microsoft Excel}
\description{
The excel.link package mainly consists of two rather independent parts: one
is for transferring data/graphics to running instance of Excel, another part
- work with data table in Excel in similar way as with usual data.frame.
}

