% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_row}
\alias{index_col}
\alias{index_row}
\alias{match_col}
\alias{match_row}
\title{Match}
\usage{
match_row(criterion, ...)

match_col(criterion, ...)

index_row(index, ...)

index_col(index, ...)
}
\arguments{
\item{criterion}{}

\item{...}{}
}
\value{
vector
}
\description{
Match
}
\examples{

a = 1
}

