% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{modify}
\alias{modify}
\alias{modify_if}
\title{Modify data.frame}
\usage{
modify(data, expr, ...)

modify_if(data, cond, expr, ...)
}
\arguments{
\item{data}{}

\item{expr}{}

\item{...}{}
}
\value{

}
\description{
Modify data.frame
}
\examples{
a = 1
}

