% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{name_dots}
\alias{name_dots}
\title{Bug workaround}
\usage{
name_dots(...)
}
\arguments{
\item{...}{arguments}
}
\value{
list
}
\description{
Function is added to workaround strange bug with data.table (issue #10).
}
