% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nest}
\alias{\%nest\%}
\alias{nest}
\title{Compute nested variable from several variables}
\usage{
nest(...)

x \%nest\% y
}
\arguments{
\item{...}{vectors/data.frames/lists}

\item{x}{vector/data.frame/list}

\item{y}{vector/data.frame/list}
}
\value{
vector/data.frame/list
}
\description{
\code{nest} mainly intended for usage with table functions such as 
\link{cro}. See examples. \code{\%nest\%} is infix version of this function.
\code{data.frame} among arguments is considered as multiple-response variable.
}
\examples{
data(mtcars)

mtcars = apply_labels(mtcars,
                      cyl = "Number of cylinders",
                      vs = "Engine",
                      vs = ml_left("
                             0 V-engine 
                             1 Straight engine
                             "),
                      am = "Transmission",
                      am = ml_left("
                             0 Automatic 
                             1 Manual
                             "),
                      carb = "Number of carburetors"
)

with(mtcars, cro(cyl, am \%nest\% vs))

# three variables 
with(mtcars, cro(am \%nest\% vs \%nest\% carb, cyl))

# the same with usual version
with(mtcars, cro(cyl, nest(am, vs)))

# three variables 
with(mtcars, cro(nest(am, vs, carb), cyl))
}
\seealso{
See also \link[base]{interaction}
}

