% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepend_names.R
\name{prepend_values}
\alias{prepend_values}
\alias{prepend_names}
\alias{prepend_all}
\alias{tab_prepend_values}
\alias{tab_prepend_names}
\alias{tab_prepend_all}
\title{Prepend values/variable names to value labels/variable labels}
\usage{
prepend_values(x)

prepend_names(x)

prepend_all(x)

tab_prepend_values(data)

tab_prepend_names(data)

tab_prepend_all(data)
}
\arguments{
\item{x}{vector/data.frame. \code{prepend_names} can be applied only to data.frames.}

\item{data}{data.frame/intermediate result of tables construction. See \link{tables}.}
}
\value{
original object with prepended names/values to labels
}
\description{
These functions add values/variable names as prefix to value labels/variable names.
Functions which starts with \code{tab_} intended for usage with tables. See \link{tables}.
}
\examples{
data(mtcars)
mtcars = apply_labels(mtcars,
                      mpg = "Miles/(US) gallon",
                      cyl = "Number of cylinders",
                      disp = "Displacement (cu.in.)",
                      hp = "Gross horsepower",
                      drat = "Rear axle ratio",
                      wt = "Weight (lb/1000)",
                      qsec = "1/4 mile time",
                      vs = "Engine",
                      vs = c("V-engine" = 0,
                             "Straight engine" = 1),
                      am = "Transmission",
                      am = c("Automatic" = 0,
                             "Manual"=1),
                      gear = "Number of forward gears",
                      carb = "Number of carburetors"
)

mtcars \%>\% prepend_names
       \%>\% calculate(
              cro_cpct(list(cyl, gear), list(total(), vs, am))
           )
}
