% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc}
\alias{qc}
\title{Create vector of characters from unquoted strings (variable names)}
\usage{
qc(...)
}
\arguments{
\item{...}{unquoted names of variables}
}
\value{
Vector of characters
}
\description{
In many cases one need to address variables in list/data.frame in such 
manner: \code{dfs[ , c("var1", "var2", "var3")]}. \code{qc} ("quoted c") is a
shortcut for the such cases to reduce keystrokes. With \code{qc} you can write:
\code{dfs[ , qc(var1, var2, var3)]}.
}
\examples{
qc(a, b, c)
identical(qc(a, b, c), c("a", "b", "c"))

mtcars[, qc(am, mpg, gear)]

}

