% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expss.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{\%<>\%}
\alias{reexports}
\alias{\%$\%}
\alias{reexports}
\alias{data.table}
\alias{reexports}
\alias{as.data.table}
\alias{reexports}
\alias{is.data.table}
\alias{reexports}
\alias{setDF}
\alias{reexports}
\alias{setDT}
\alias{reexports}
\alias{setkey}
\alias{reexports}
\alias{setkeyv}
\alias{reexports}
\alias{fread}
\alias{reexports}
\alias{fwrite}
\alias{reexports}
\alias{htmlTable}
\alias{reexports}
\alias{htmlTableWidget}
\alias{reexports}
\alias{htmlTableWidgetOutput}
\alias{reexports}
\alias{interactiveTable}
\alias{reexports}
\alias{dataTableAjax}
\alias{reexports}
\alias{dataTableOutput}
\alias{reexports}
\alias{renderDataTable}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{data.table}}, \code{\link[data.table]{as.data.table}}, \code{\link[data.table]{is.data.table}}, \code{\link[data.table]{setDF}}, \code{\link[data.table]{setDT}}, \code{\link[data.table]{setkey}}, \code{\link[data.table]{setkeyv}}, \code{\link[data.table]{fread}}, \code{\link[data.table]{fwrite}}}

  \item{DT}{\code{\link[DT]{dataTableAjax}}, \code{\link[DT]{dataTableOutput}}, \code{\link[DT]{renderDataTable}}}

  \item{htmlTable}{\code{\link[htmlTable]{htmlTable}}, \code{\link[htmlTable]{htmlTableWidget}}, \code{\link[htmlTable]{htmlTableWidgetOutput}}, \code{\link[htmlTable]{interactiveTable}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%<>\%}}, \code{\link[magrittr]{\%$\%}}}
}}

