% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_cell_chisq.R
\name{significance_cell_chisq}
\alias{significance_cell_chisq}
\alias{cell_chisq}
\title{Compute cell chi-square test on table with column percents}
\usage{
significance_cell_chisq(x, sig_level = 0.05, min_base = 2,
  subtable_marks = c("both", "greater", "less"), sig_labels = c("<",
  ">"), correct = TRUE, keep = c("percent", "bases", "none"),
  row_margin = c("auto", "sum_row", "first_column"),
  total_marker = "#", total_row = 1, total_column_marker = "#",
  digits = get_expss_digits())

cell_chisq(cases_matrix, row_base, col_base, total_base, correct)
}
\arguments{
\item{x}{table (class \code{etable}): result of \link{cro_cpct} with 
proportions and bases.}

\item{sig_level}{numeric. Significance level - by default it equals to \code{0.05}.}

\item{min_base}{numeric. Significance test will be conducted if 
column have base greater or equal to \code{min_base}. By default it equals to \code{2}.}

\item{subtable_marks}{character. One of "both" , "greater" or "less". By
default it is "both".}

\item{sig_labels}{a character vector with two elements. Labels 
for marking difference with row margin of the table. First mark means
'lower' (by default it is \code{<}) and the second means 'greater'
(\code{>}).}

\item{correct}{logical indicating whether to apply continuity correction when
computing the test statistic for 2 by 2 tables. For details see
\link[stats]{chisq.test}.}

\item{keep}{character. One or more from "percent", "bases", 
or "none". This argument determines which statistics will remain in
the table after significance marking.}

\item{row_margin}{character. One of values "auto", "sum_row", or "first_column"}

\item{total_marker}{character. Mark of total rows in the table. "#" by default.}

\item{total_row}{integer/character. In case of several totals per subtable it
is number or name of total row for significance calculation.}

\item{total_column_marker}{character}

\item{digits}{an integer indicating how much digits after decimal separator 
will be shown in the final table.}

\item{cases_matrix}{matrix with counts size R*C}

\item{row_base}{vector with row bases, length R}

\item{col_base}{vector with col bases, length C}

\item{total_base}{single value, total base}
}
\description{
Compute cell chi-square test on table with column percents
}
