% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_cases.R, R/significance_cpct.R,
%   R/significance_means.R
\name{significance_cases}
\alias{significance_cases}
\alias{significance_cpct}
\alias{add_sig_labels}
\alias{significance_means}
\title{Mark significant differences between columns of the table}
\usage{
significance_cases(x, sig_level = 0.05, min_base = 2, keep = c("cases",
  "bases"), total_marker = "#", total_row = 1,
  digits = get_expss_digits())

significance_cpct(x, sig_level = 0.05, delta_cpct = 0, min_base = 2,
  compare_type = "subtable", bonferroni = FALSE,
  subtable_marks = c("greater", "both", "less"), inequality_sign = "both"
  \%in\% subtable_marks, sig_labels = LETTERS,
  sig_labels_previous_column = c("v", "^"), sig_labels_first_column = c("-",
  "+"), keep = c("percent", "bases"), na_as_zero = FALSE,
  total_marker = "#", total_row = 1, digits = get_expss_digits())

add_sig_labels(tbl, sig_labels = LETTERS)

significance_means(x, sig_level = 0.05, delta_means = 0, min_base = 2,
  compare_type = "subtable", bonferroni = FALSE,
  subtable_marks = c("greater", "both", "less"), inequality_sign = "both"
  \%in\% subtable_marks, sig_labels = LETTERS,
  sig_labels_previous_column = c("v", "^"), sig_labels_first_column = c("-",
  "+"), keep = c("means", "sd", "bases"), var_equal = FALSE,
  digits = get_expss_digits())
}
\arguments{
\item{x}{table (class \code{etable}): result of \link{cro_cpct} with 
proportions and bases for \code{significance_cpct}, result of 
\link{cro_mean_sd_n} with means, standard deviations and valid N for 
\code{significance_means}, and result of \link{cro_cases} with counts and
bases for \code{significance_cases}.}

\item{sig_level}{numeric. Significance level - by default it equals to \code{0.05}.}

\item{min_base}{numeric. Significance test will be conducted if both
columns have bases greater than \code{min_base}. By default it equals to \code{2}.}

\item{keep}{character. One or more from "percent", "cases", "means", "bases", 
"sd" or "none". This argument determines which statistics will remain in
the table after significance marking.}

\item{total_marker}{character. Mark of total rows in table.}

\item{total_row}{integer/character. In case of several totals per subtable it is
number or name of total row for significance calculation.}

\item{digits}{an integer indicating how much digits after decimal separator
will be shown in the final table.}

\item{delta_cpct}{numeric. Minimal delta between percent for which we mark 
significant differences (in percent points) - by default it equals to zero.
Note that, for example, for minimal 5 percent difference
\code{delta_cpct} should be equals 5, not 0.05.}

\item{compare_type}{Type of compare between columns. By default it is 
\code{subtable} - comparisons will be conducted between columns of each 
subtable. Other possible values are: \code{first_column}, 
\code{adjusted_first_column} and \code{previous_column}. We can conduct
several tests simultaneously.}

\item{bonferroni}{logical. \code{FALSE} by default. Should we use Bonferroni
adjustment by number of comparisons in each row?}

\item{subtable_marks}{character. One of "greater", "both" or "less". By
deafult we mark only values which are significantly greater than some other
columns. We can change this behavior by setting argument to \code{less} or
\code{both}.}

\item{inequality_sign}{logical. FALSE if \code{subtable_marks} is "less" or 
"greater". Should we show \code{>} or \code{<} before significance marks of
subtable comparisons.}

\item{sig_labels}{character vector. Labels for marking differences between
columns of subtable.}

\item{sig_labels_previous_column}{a character vector with two elements. Labels
for marking difference with previous column. First mark means 'lower' (by
default it is \code{v}) and the second means greater (\code{^}).}

\item{sig_labels_first_column}{a character vector with two elements. Labels
for marking difference with first column of the table. First mark means 'lower' (by
default it is \code{-}) and the second means 'greater' (\code{+}).}

\item{na_as_zero}{logical. \code{FALSE} by default. Should we treat
\code{NA}'s as zero cases?}

\item{delta_means}{numeric. Minimal delta between means for which we mark 
significant differences  - by default it equals to zero.}

\item{var_equal}{a logical variable indicating whether to treat the two
variances as being equal. For details see \link[stats]{t.test}.}
}
\value{
Object of class \code{etable} with marks of significant differences
  between columns.
}
\description{
\itemize{
\item{\code{significance_cpct}}{ conducts z-tests between column percent in
the result of \link{cro_cpct}. Results are calculated with the same formula 
as in \link[stats]{prop.test} without continuity correction.}
\item{\code{significance_means}}{ conducts t-tests between column means in
the result of \link{cro_mean_sd_n}. Results are calculated with the same formula 
as in \link[stats]{t.test}.}
\item{\code{significance_cases}}{ conducts chi-squared tests on the subtable of
table with counts in the result of \link{cro_cases}. Results are calculated
with the same formula as in \link[stats]{chisq.test} without continuity
correction.}} 
There are four type of comparisons which can be conducted simultaneously 
(argument \code{compare_type}). \code{subtable} provide comparison between
all columns inside each subtable. \code{previous_column} is comparison of
each column in the subtable with previous column. It is useful if columns are
periods or waves of survey. \code{first_column} provide comparison of table
first column with all other columns in the table.
\code{adjusted_first_column} is comparison with first column but with
adjustment for common base. It is useful if first column is total column and
other columns are subgroup of this total. Adjustments are made according to
algorithm in IBM SPSS Statistics Algorithms v20, p. 263. Note that with these
adjustments t-tests between means are made with equal variance assumed (as
with \code{var_equal = TRUE}).
}
\examples{
data(mtcars)
mtcars = apply_labels(mtcars,
                      mpg = "Miles/(US) gallon",
                      cyl = "Number of cylinders",
                      disp = "Displacement (cu.in.)",
                      hp = "Gross horsepower",
                      drat = "Rear axle ratio",
                      wt = "Weight (lb/1000)",
                      qsec = "1/4 mile time",
                      vs = "Engine",
                      vs = c("V-engine" = 0,
                             "Straight engine" = 1),
                      am = "Transmission",
                      am = c("Automatic" = 0,
                             "Manual"=1),
                      gear = "Number of forward gears",
                      carb = "Number of carburetors"
)

mtcars_table = calculate(mtcars,
                   cro_cpct(list(cyl, gear),
                            list(total(), vs, am))
                         )

significance_cpct(mtcars_table)

# comparison with first column
significance_cpct(mtcars_table, compare_type = "first_column")

# comparison with first column and inside subtable
significance_cpct(mtcars_table, 
            compare_type = c("first_column", "subtable"))

# only significance marks
significance_cpct(mtcars_table, keep = "none")

# means
mtcars_means = calculate(mtcars,
                   cro_mean_sd_n(list(mpg, wt, hp),
                                 list(total(), vs, cyl))
                        )
                        
significance_means(mtcars_means) 

# mark values which are less and greater
significance_means(mtcars_means, subtable_marks = "both")

# chi-squared test
mtcars_cases = calculate(mtcars,
                   cro_cases(list(cyl, gear),
                            list(total(), vs, am))
                         )
                         
significance_cases(mtcars_cases)                         
}
\seealso{
\link{cro_cpct}, \link{cro_cases}, \link{cro_mean_sd_n}, 
  \link{compare_proportions}, \link{compare_means}, \link[stats]{prop.test},
  \link[stats]{t.test}, \link[stats]{chisq.test}
}
