% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_labels.R
\name{split_labels}
\alias{split_labels}
\alias{split_columns}
\title{Split character vector to matrix/split columns in data.frame}
\usage{
split_labels(x, remove_repeated = TRUE, split = "|", fixed = TRUE,
  perl = FALSE)

split_columns(data, columns = 1, remove_repeated = TRUE, split = "|",
  fixed = TRUE, perl = FALSE)
}
\arguments{
\item{x}{character vector which will be split}

\item{remove_repeated}{logical. Should we remove repeated labels?}

\item{split}{character vector (or object which can be coerced to such)
containing regular expression(s) (unless \code{fixed = TRUE}) to use for
splitting.}

\item{fixed}{logical. If TRUE match split exactly, otherwise use regular
expressions. Has priority over \code{perl}.}

\item{perl}{logical. Should Perl-compatible regexps be used?}

\item{data}{data.frame vector which will be split}

\item{columns}{character/numeric/logical  columns in the data.frame
\code{data} which should be split}
}
\value{
\code{split_labels} returns character matrix, \code{split_columns} returns
  data.frame with columns replaced by possibly multiple columns with split
  labels.
}
\description{
\code{split_labels}/\code{split_columns} are auxiliary functions for 
post-processing tables resulted from \link{table_cases}/\link{table_summary} 
and etc. In these tables all labels collapsed in the first column with "|" 
separator. \code{split_columns} split first column into multiple columns 
with separator (\code{split} argument).
}
\examples{
data(mtcars)

# apply labels
mtcars = apply_labels(mtcars,
    cyl = "Number of cylinders",
    vs = "Engine",
    vs = c("V-engine" = 0,
                    "Straight engine" = 1),
    am = "Transmission",
    am = c(automatic = 0,
                    manual=1),
    gear = "Number of forward gears",
    carb = "Number of carburetors"
)

# all row labels in the first column
tabl = mtcars \%>\% calculate(table_cpct(list(cyl, gear, carb), list("#total", vs, am)))
split_labels(tabl[[1]])
split_labels(colnames(tabl))

# replace first column with new columns 
split_columns(tabl) # remove repeated

split_columns(tabl, remove_repeated = FALSE)
}
\seealso{
\link[base]{strsplit}
}
