% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_separate.R
\name{split_separate}
\alias{split_separate}
\alias{split_off}
\title{Splits data.frame into list of data.frames that can be analyzed separately}
\usage{
split_separate(data, ..., drop = TRUE)

split_off(data, add_group = FALSE, add_rownames = FALSE)
}
\arguments{
\item{data}{data.frame for \code{split_separate}/list for \code{split_off}}

\item{...}{unquoted variables names (see \link{keep}) by which \code{data}
will be split into list.}

\item{drop}{should we drop combination of levels with zero observation? TRUE
by default.}

\item{add_group}{Should we add group values as \code{.group} variable to
result of \code{split_off}? FALSE by default.}

\item{add_rownames}{Should we add rownames as \code{.rownames} variable to
result of \code{split_off}? FALSE by default.}
}
\value{
\code{split_separate} returns list of data.frames/\code{split_off}
  returns data.frame
}
\description{
Splits data.frame into list of data.frames that can be analyzed separately. 
These data.frames are sets of cases that have the same values for the 
specified split variables. Any missing values in split variables are dropped 
together with the corresponding values of \code{data}. \code{split_off} works
only with lists of data frames (assumed to have compatible structure).
Resulting rows will be sorted in order of the split variables.
}
\examples{

# example from base R 'split'
data(airquality)
airquality2 = airquality \%>\% 
    split_separate(Month) \%>\% 
    compute({
        Ozone_zscore = scale(Ozone)
    }) \%>\% 
    split_off() 
head(airquality2)
        
}
\seealso{
\link[base]{split}, \link{compute}, \link{calculate},
  \link{do_repeat}, \link{where}
}
