% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{sum_row}
\alias{apply_col}
\alias{apply_row}
\alias{max_col}
\alias{max_row}
\alias{mean_col}
\alias{mean_row}
\alias{median_col}
\alias{median_row}
\alias{min_col}
\alias{min_row}
\alias{sd_col}
\alias{sd_row}
\alias{sum_col}
\alias{sum_row}
\title{Sum/mean/sd/median/max/min on row/column}
\usage{
sum_row(...)

sum_col(...)

mean_row(...)

mean_col(...)

sd_row(...)

sd_col(...)

median_row(...)

median_col(...)

max_row(...)

max_col(...)

min_row(...)

min_col(...)

apply_row(fun, ...)

apply_col(fun, ...)
}
\arguments{
\item{...}{data. Vectors, matrixes, data.frames, list. Shorter arguments
will be recycled.}

\item{criterion}{Vector with counted values, list with conditions or
function. If criterion is missing (or is NULL) non-NA's values will be
counted.}

\item{x}{Counted values or criterion for counting. Vector, matrix, data.frame,
list, function. Shorter columns in list will be recycled.}
}
\value{
\code{count} return single value (vector of length 1). 
\code{count_row} returns vector of counts for each row of supplied arguments.
\code{count_col} returns vector of counts for each column of supplied arguments.
\code{\%has\%} returns logical vector - presence indicator of criterion in each row.
}
\description{
There are two flavors of this function - one works with entire dataset/matrix/vector
similar to Microsoft Excel \code{COUNTIF}. The second works rowwise - e. g. 
similar to SPSS \code{COUNT} function.
}
\details{
\code{count} counts values in entire dataset and return single 
value (vector of length 1).

\code{count_row} counts values in each row of supplied arguments and return
vector of counts for each row of supplied arguments.

\code{count_col} counts values in each column of supplied arguments and return
vector of counts for each column of supplied arguments.

All functions never return NA's. 

If criterion is list, then each element considered as single condition and
they will be combined with AND.

Function criterion should return logicals of same size and shape as its argument.
This function will be applied to each column of supplied data and TRUE results will be counted.
There is asymmetrical behavior in \code{count_row} and
\code{count_col}: in both cases function criterion will be applied
columnwise.
There are special functions for usage as criteria (e. g. \code{gt(5)} is
equivalent ">5" in spreadsheet):
\itemize{
\item{\code{gt}}{ greater than}
\item{\code{gte}}{ greater than or equal}
\item{\code{eq}}{ equal} 
\item{\code{neq}}{ not equal} 
\item{\code{lt}}{ less than}
\item{\code{lte}}{ less than or equal}
}
}
\examples{
# Examples borrowed from Microsoft Excel help for COUNTIF
df1 = data.frame(
    a=c("apples",   "oranges",     "peaches",     "apples"),
    b = c(32, 54, 75, 86)
)

count("apples",df1$a) # 2

count("apples",df1) # 2

with(df1,count("apples",a,b)) # 2

count(gt(55),df1$b) # greater than 55 = 2

count(neq(75),df1$b) # not equal 75 = 3

count(gte(32),df1$b) # greater than or equal 32 = 4

count(list(gt(32), lt(86)),df1$b) # 2

count(gt(32) & lt(86),df1$b) # 2

count(33:85,df1$b) # 2

# more complex criteria
# values with letters
count(function(x) grepl("^[A-z]+$",x),df1) # 4

# values that started on 'a'
count(function(x) grepl("^a",x),df1) # 2

# count_row
count_row(function(x) grepl("^a",x),df1) # c(1,0,0,1)

df1 \%has\% 'apples' # c(TRUE,FALSE,FALSE,TRUE)

# example with dplyr
if (require(dplyr)){
 set.seed(123)
 df2 = as.data.frame(
        matrix(sample(c(1:10,NA),30,replace = TRUE),10)
 )
 df2  \%>\% mutate(exact = count_row(8, V1, V2, V3),
                    greater = count_row(gt(8), V1, V2, V3),
                    range = count_row(5:8, V1, V2, V3),
                    na = count_row(is.na, V1, V2, V3),
                    not_na = count_row(, V1, V2, V3)
                 ) -> result
 result
}
}

