% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tables.R
\name{tab_cols}
\alias{tab_cols}
\alias{tab_cells}
\alias{tab_rows}
\alias{tab_stat_fun}
\alias{tab_stat_fun_df}
\alias{tab_stat_cases}
\alias{tab_stat_cpct}
\alias{tab_stat_cpct_responses}
\alias{tab_stat_tpct}
\alias{tab_weight}
\alias{tab_subgroup}
\title{Functions for constracting custom tables}
\usage{
tab_cols(data, ...)

tab_cells(data, ...)

tab_rows(data, ...)

tab_stat_fun(data, ..., label = NULL)

tab_stat_fun_df(data, ..., label = NULL)

tab_stat_cases(data, label = NULL)

tab_stat_cpct(data, label = NULL)

tab_stat_cpct_responses(data, label = NULL)

tab_stat_tpct(data, label = NULL)

tab_weight(data, weight = NULL)

tab_subgroup(data, subgroup = NULL)
}
\arguments{
\item{...}{jjjkj}
}
\value{
jkjk
}
\description{
Functions for constracting custom tables
}
\examples{
1
}
