% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tables_significance.R
\name{tab_significance_options}
\alias{tab_significance_options}
\alias{tab_stat_cpct_significance}
\alias{tab_last_round}
\alias{tab_last_add_sig_labels}
\title{Significance testing for custom tables}
\usage{
tab_significance_options(data, sig_level = 0.05, min_base = 2,
  delta_cpct = 0, delta_means = 0, compare_type = "subtable",
  bonferroni = FALSE, subtable_marks = "greater", inequality_sign = "both"
  \%in\% subtable_marks, sig_labels = LETTERS,
  sig_labels_previous_column = c("v", "^"), sig_labels_first_column = c("-",
  "+"), keep = c("percent", "cases", "means", "sd", "bases"),
  total_marker = "#", total_row = 1, digits = get_expss_digits(),
  na_as_zero = FALSE, var_equal = FALSE)

tab_stat_cpct_significance(data, sig_level = 0.05, delta_cpct = 0,
  min_base = 2, compare_type = "subtable", bonferroni = FALSE,
  subtable_marks = c("greater", "both", "less"), inequality_sign = "both"
  \%in\% subtable_marks, sig_labels = LETTERS,
  sig_labels_previous_column = c("v", "^"), sig_labels_first_column = c("-",
  "+"), keep = c("percent", "bases"), na_as_zero = FALSE,
  total_marker = "#", total_row = 1, digits = get_expss_digits(),
  label = NULL)

tab_last_round(data, digits = get_expss_digits())

tab_last_add_sig_labels(data, sig_labels = LETTERS)
}
\arguments{
\item{data}{data.frame/intermediate_table}

\item{sig_level}{numeric. Significance level - by default it equals to \code{0.05}.}

\item{min_base}{numeric. Significance test will be conducted if both
columns have bases greater than \code{min_base}. By default it equals to \code{2}.}

\item{delta_cpct}{numeric. Minimal delta between percent for which we mark 
significant differences (in percent points) - by default it equals to zero.
Note that, for example, for minimal 5 percent difference
\code{delta_cpct} should be equals 5, not 0.05.}

\item{delta_means}{numeric. Minimal delta between means for which we mark 
significant differences  - by default it equals to zero.}

\item{compare_type}{Type of compare between columns. By default it is 
\code{subtable} - comparisons will be conducted between columns of each 
subtable. Other possible values are: \code{first_column}, 
\code{adjusted_first_column} and \code{previous_column}. We can conduct
several tests simultaneously.}

\item{bonferroni}{logical. \code{FALSE} by default. Should we use Bonferroni
adjustment by number of comparisons in each row?}

\item{subtable_marks}{character. One of "greater", "both" or "less". By
deafult we mark only values which are significantly greater than some other
columns. We can change this behavior by setting argument to \code{less} or
\code{both}.}

\item{inequality_sign}{logical. FALSE if \code{subtable_marks} is "less" or 
"greater". Should we show \code{>} or \code{<} before significance marks of
subtable comparisons.}

\item{sig_labels}{character vector. Labels for marking differences between
columns of subtable.}

\item{sig_labels_previous_column}{a character vector with two elements. Labels
for marking difference with previous column. First mark means 'lower' (by
default it is \code{v}) and the second means greater (\code{^}).}

\item{sig_labels_first_column}{a character vector with two elements. Labels
for marking difference with first column of the table. First mark means 'lower' (by
default it is \code{-}) and the second means 'greater' (\code{+}).}

\item{keep}{character. One or more from "percent", "cases", "means", "bases", 
"sd" or "none". This argument determines which statistics will remain in
the table after significance marking.}

\item{total_marker}{character. Mark of total rows in table.}

\item{total_row}{integer/character. In case of several totals per subtable it is
number or name of total row for significance calculation.}

\item{digits}{an integer indicating how much digits after decimal separator
will be shown in the final table.}

\item{na_as_zero}{logical. \code{FALSE} by default. Should we treat
\code{NA}'s as zero cases?}

\item{var_equal}{a logical variable indicating whether to treat the two
variances as being equal. For details see \link[stats]{t.test}.}
}
\value{
All of these functions return object of class 
  \code{intermediate_table}. Use \code{tab_pivot} to get final result -
  object of class \code{etable}.
}
\description{
Significance testing for custom tables
}
