% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_cases.R
\name{table_cases}
\alias{table_cases}
\alias{table_cpct}
\alias{table_rpct}
\alias{table_tpct}
\title{Title}
\usage{
table_cases(row_vars, col_vars, weight = NULL,
  total_row_position = c("below", "above", "none"),
  total_row_title = "#Total", weighted_total = FALSE,
  variable_labels_position = c("rows", "inside_columns", "outside_columns"))

table_cpct(row_vars, col_vars, weight = NULL,
  total_row_position = c("below", "above", "none"),
  total_row_title = "#Total", weighted_total = FALSE,
  variable_labels_position = c("rows", "inside_columns", "outside_columns"))

table_rpct(row_vars, col_vars, weight = NULL,
  total_row_position = c("below", "above", "none"),
  total_row_title = "#Total", weighted_total = FALSE,
  variable_labels_position = c("rows", "inside_columns", "outside_columns"))

table_tpct(row_vars, col_vars, weight = NULL,
  total_row_position = c("below", "above", "none"),
  total_row_title = "#Total", weighted_total = FALSE,
  variable_labels_position = c("rows", "inside_columns", "outside_columns"))
}
\arguments{
\item{row_vars}{sdds}

\item{col_vars}{sdsd}

\item{weight}{sddssd}

\item{total_row_position}{sdsdds}

\item{total_row_title}{sdsdd}

\item{weighted_total}{dssds}

\item{variable_labels_position}{sddsds}
}
\value{
data.frame
}
\description{
Title
}
\examples{

data(mtcars)

# apply labels
mtcars = modify(mtcars,{
    var_lab(cyl) = "Number of cylinders"
    var_lab(vs) = "Engine"
    val_lab(vs) = c("V-engine" = 0,
                    "Straight engine" = 1)
    var_lab(am) = "Transmission"
    val_lab(am) = c(automatic = 0,
                    manual=1)
    var_lab(gear) = "Number of forward gears"
    var_lab(carb) = "Number of carburetors"
})

# simple frequency table
table_cases(mtcars$vs, mtcars$am)

# Total column, several col_vars and variables, column percent
with(mtcars,
     table_cpct(list(cyl, gear, carb), list("#total", vs, am))
     )

# the same table without total column and total row
with(mtcars,
     table_cpct(list(cyl, gear, carb), list(vs, am), total_row_position = "none")
     )

}
