% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_summary.R
\name{table_summary}
\alias{table_summary}
\alias{table_means}
\alias{table_medians}
\alias{table_sums}
\title{Title}
\usage{
table_summary(summary_vars, col_vars, fun, weight = NULL, row_vars = NULL,
  stat_names = NULL, row_labels = c("row_vars", "row_vars_values",
  "summary_vars", "stat_names"), col_labels = c("col_vars",
  "col_vars_values"), hide = NULL)

table_means(summary_vars, col_vars, weight = NULL, row_vars = NULL,
  stat_names = c("Mean", "Std. Dev.", "N"))

table_medians(summary_vars, col_vars, weight = NULL, row_vars = NULL,
  stat_names = c("Median", "N"))

table_sums(summary_vars, col_vars, weight = NULL, row_vars = NULL,
  stat_names = c("Sum", "N"))
}
\arguments{
\item{summary_vars}{vector/data.frame/list - variables for summary. If ommited
then variables will be used as summary variables.}

\item{col_vars}{fss}

\item{fun}{the function to be applied to summary variables}

\item{weight}{dsvfv}

\item{row_vars}{fdfg}

\item{stat_names}{cdsdc}

\item{row_labels}{csdsd}

\item{col_labels}{dcdscds}

\item{hide}{dcsdc}
}
\value{
table
}
\description{
Title
}
\examples{
a = 1
}
