% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unlab}
\alias{unlab}
\title{Drop variable label and value labels}
\usage{
unlab(x)
}
\arguments{
\item{x}{Variable(s). Vector/data.frame/list.}
}
\value{
\code{unlab} returns original variable x without variable label and value labels.
}
\description{
\code{unlab} returns variable x without variable labels and value labels
}
\examples{
raw_var = rep(1:2,5)
var_with_lab = set_var_lab(raw_var,"Income")
val_lab(var_with_lab) = c("Low"=1,"High"=2)
stopifnot(identical(raw_var,unlab(var_with_lab)))
}
\seealso{
\code{link{unvr}} \code{link{unvl}}
}

