% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/labels.R
\name{val_lab}
\alias{add_val_lab<-}
\alias{make_labels}
\alias{set_val_lab}
\alias{unvl}
\alias{val_lab}
\alias{val_lab<-}
\title{Set or get value labels}
\usage{
val_lab(x)

val_lab(x) <- value

set_val_lab(x, value, add = FALSE)

add_val_lab(x) <- value

make_labels(text, code_position = c("left", "right"))
}
\arguments{
\item{x}{Variable(s). Vector/data.frame/list.}

\item{value}{Named vector. Names of vector values are labels for the
appropriate values of variable x.}

\item{add}{Logical. Should value labels replace old labels? Deafult is FALSE. If TRUE
new value lables will be combined with old value labels.}

\item{text}{text that should be converted to named vector}

\item{code_position}{Possible values "left" or "right" - position of numeric code in
\code{text}.}
}
\value{
\code{val_lab} return value labels (named vector). If labels doesn't
  exist it return NULL . \code{val_lab<-} and \code{set_val_lab} return
  variable (vector x) of class "with_labels" with attribute "value_labels"
  which contains value labels. \code{make_labels} return named vector for usage as value labels.
}
\description{
\code{val_lab} returns value labels or NULL if labels doesn't
exist.
\code{val_lab<-} set value labels.
\code{set_val_lab} returns variable with value labels.
\code{add_val_lab<-} add value labels to already existing value labels.
\code{unvl} drops value labels.
\code{make_labels} return named vector for usage as value labels.
}
\details{
Value labels are stored in attribute "value_labels"
  (\code{attr(x,"value_labels")}). Duplicated values are not allowed. If
  argument \code{x} is data.frame or list then labels applied to all elements
  of data.frame/list. We set variable class to "with_labels" for preserving
  labels from dropping during some operations (such as \code{c} and \code{[}).
  There are special methods of subsetting and concatenation for this class.
  To drop value labels use \code{val_lab(var) <- NULL} or \code{unvl(var)}.
  \code{make_labels} converts text in the form that used in questionnaires
  to named vector. See examples.
}
\examples{
data(ProductTest)

### Common usage ###
val_lab(ProductTest$s2b) = c('18 - 26' = 2, '27 - 35' = 3)

head(factor(ProductTest$s2b))

identical(levels(factor(ProductTest$s2b)), names(val_lab(ProductTest$s2b)))

ProductTest = unlab(ProductTest) # drop all labels

### Add labels ###

age_groups = c('18 - 26' = 2, '27 - 35' = 3)

add_val_lab(ProductTest$s2b) = age_groups[1]
add_val_lab(ProductTest$s2b) = age_groups[2]

identical(ProductTest$s2b, set_val_lab(ProductTest$s2b,age_groups))

ProductTest$s2b = unlab(ProductTest$s2b)

## make labels from text copied from questionnaire

val_lab(ProductTest$s2b) = make_labels("
 1. 18 - 26
 2. 27 - 35
")

head(factor(ProductTest$s2b))

# or, if in original codes is on the right side

val_lab(ProductTest$s8_1) = make_labels("
 Chocolate bars    1
 Chocolate sweets (bulk)	2
 Slab chocolate(packed)	3
 Slab chocolate (bulk)	4
 Boxed chocolate sweets	5
 Marshmallow/pastilles in chocolate coating	6
 Marmalade in chocolate coating	7
 Other	8
", code_position = "right")

head(factor(ProductTest$s8_1))
}

