% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values2labels.R
\name{values2labels}
\alias{v2l}
\alias{values2labels}
\title{Replace vector/matrix/data.frame/list values with corresponding value labels.}
\usage{
values2labels(x)

v2l(x)
}
\arguments{
\item{x}{vector/matrix/data.frame/list}
}
\value{
Object of the same form as x but with value labels instead of values. If there are no labels for
 some values they are converted to characters in most cases.
}
\description{
Replace vector/matrix/data.frame/list values with corresponding value labels.
}
\details{
\code{v2l} is just shortcut to \code{values2labels}.
}
\examples{
data(mtcars)
mtcars = within(mtcars,{
                var_lab(mpg) = NULL
                val_lab(am) = c(" automatic" = 0, " manual" =  1)
})

summary(lm(mpg ~ ., data = values2labels(mtcars[,c("mpg","am")])))
}
\seealso{
\code{\link{names2labels}}
}

