% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/labels.R
\name{var_lab}
\alias{set_var_lab}
\alias{unvr}
\alias{var_lab}
\alias{var_lab<-}
\title{Set or get variable label}
\usage{
var_lab(x)

var_lab(x) <- value

set_var_lab(x, value)

unvr(x)
}
\arguments{
\item{x}{Variable. In the most cases it is numeric vector.}

\item{value}{A character scalar - label for the variable x.}
}
\value{
\code{var_lab} return variable label. If label doesn't exist it return
  NULL . \code{var_lab<-} and \code{set_var_lab} return variable (vector x)
  of class "with_labels" with attribute "label" which equals submitted value.
}
\description{
\code{var_lab} returns variable label or NULL if label doesn't
exist.
\code{var_lab<-} set variable label.
\code{set_var_lab} returns variable with label.
\code{unvr} drops variable label.
}
\details{
Variable label is stored in attribute "label" (\code{attr(x,"label")}). For
  preserving from droping this attribute during some operations (such as \code{c})
  we set variable class to "with_labels". There are special methods of
  subsetting and concatenation for this class. To drop variable label use
  \code{var_lab(var) <- NULL} or \code{unvr(var)}.
}
\examples{
a = rep(1:2,5)
b = a
var_lab(a) = "Gender"
b = set_var_lab(b,"Gender")
identical(a,b) # should be TRUE

identical(var_lab(a),attr(a,"label")) # should be TRUE
}

