% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{var_range}
\alias{\%to\%}
\alias{var_range}
\title{Get range of variables/variables by pattern}
\usage{
var_range(start, end)

e1 \%to\% e2
}
\arguments{
\item{start}{character Name of start variable (e. g. a_1)}

\item{end}{character Name of start variable (e. g. a_5)}

\item{e1}{unquoted name of start variable (e. g. a_1)}

\item{e2}{unquoted name of start variable (e. g. a_5)}
}
\value{
list of variables
}
\description{
Get range of variables/variables by pattern
}
\examples{
a = 1
}

