% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_lm.R
\name{w_lm}
\alias{w_lm}
\alias{summary.w_lm}
\title{Fitting linear models with frequency weights}
\usage{
w_lm(
  data,
  formula,
  weight = NULL,
  weight_is_frequency = TRUE,
  subset = NULL,
  ...
)

\method{summary}{w_lm}(object, correlation = FALSE, symbolic.cor = FALSE, ...)
}
\arguments{
\item{data}{an data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model. If
not found in data, the variables are taken from environment(formula),
typically the environment from which w_lm is called.}

\item{formula}{an object of class \link[stats]{formula} (or one that can be
coerced to that class): a symbolic description of the model to be fitted.
The details of model specification are given in \link[stats]{lm}.}

\item{weight}{an optional vector of non-negative weights to be used in the
fitting process. Should be NULL or a numeric vector. If non-NULL, weighted
least squares is used with weights weights (that is, minimizing
sum(w*e^2));}

\item{weight_is_frequency}{a logical. TRUE by default. Should we treat weight
as SPSS-style frequency weights?}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{...}{Further arguments to pass to \link[stats]{lm}.}

\item{correlation}{logical; if TRUE, the correlation matrix of the estimated
parameters is returned and printed.}

\item{symbolic.cor}{logical. If TRUE, print the correlations in a symbolic
form (see symnum) rather than as numbers.}
}
\value{
an object of class "w_lm". For details see \link[stats]{lm} and
  \link[stats]{summary.lm}.
}
\description{
\code{w_lm} is used to fit linear models with frequency weights, also known
as SPSS-style weights.  'w_lm' has 'data' as a first argument for usage with
'magrittr' pipes. Additionally, it supports variable labels. 'w_lm' uses
standard R 'lm' internally so for detailed help see \link[stats]{lm} and
\link[stats]{summary.lm}.
}
\examples{
1
}
