% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_if.R
\name{check_if}
\alias{check_if}
\title{Functions for checking conditional questions/subset of data.}
\usage{
check_if(.data, expr)
}
\arguments{
\item{.data}{data.frame for checking}

\item{expr}{Logical vector/expression.}
}
\value{
These functions returns object of class 'chk_if'. It is suited for consumption by
functions: \code{\link{sngl}},\code{\link{mult}},\code{\link{dmult}}, \code{\link{error_if}} and
\code{\link{move}}. Also it can be accepted by next \code{\link{check_if}} and/or
\code{\link{check_subset}}.
}
\description{
These functions are designed for working with piping operator
\code{\link[magrittr]{\%>\%}} from 'magrittr' package.
}
\details{
TRUE in expr vector indicated rows in .data that should contain valid
 values. For \code{check_if} in other rows all .data values should be NA. It used for questions that
  were asked by condition on answer from previous questions. For \code{check_subset} values in
  other rows will be ignored - they are all will be considered as valid.
}
\examples{
library(dplyr)
data(ProductTestRaw)
data(codeframe)

## Example 1 ##

valid_a4 = make_labels(codeframe$dislikes_in_appearance)
# question a4 was asked only if codes 1-4 marked in a3
# 3 errors: 1 missing value, 1 invalid code, 1 code in case of a3 in 5-7.
ProductTestRaw  \%>\%
     check_if(a3 \%in\% 1:4)  \%>\%
     mult(a4_1:a4_6)(valid_a4,exclusive=99)  \%>\%
     report

## Example 2 ##

# There are no errors in question s2b in cell 2.
ProductTestRaw  \%>\%
     check_subset(cell \%in\% 2) \%>\%
     sngl(s2b)(2,3)  \%>\%
     report

## Example 3 ##

# There are no errors in question a4 in cell 2.
ProductTestRaw  \%>\%
     check_subset(cell \%in\% 2)  \%>\%
     check_if(a3 \%in\% 1:4)  \%>\%
     mult(a4_1:a4_6)(valid_a4,exclusive=99)  \%>\%
     report
}
\seealso{
\code{\link{sngl}}, \code{\link{mult}},  \code{\link{dmult}},
\code{\link{error_if}} and \code{\link{move}}
}

