% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tables.R
\name{col_vars_}
\alias{col_vars_}
\alias{cell_}
\alias{row_vars_}
\alias{stat}
\alias{stat_df}
\alias{stat_}
\alias{stat_df_}
\alias{weight_}
\alias{subgroup_}
\title{Functions for constracting custom tables}
\usage{
col_vars_(..., label = NULL)

cell_(..., label = NULL)

row_vars_(..., label = NULL)

stat(..., fun_value_labels = NULL, label = NULL)

stat_df(..., fun_value_labels = NULL, label = NULL)

stat_(..., fun_value_labels = NULL, label = NULL)

stat_df_(..., fun_value_labels = NULL, label = NULL)

weight_(...)

subgroup_(..., label = NULL)
}
\arguments{
\item{...}{jjjkj}
}
\value{
jkjk
}
\description{
Functions for constracting custom tables
}
\examples{
1
}
