% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cro.R
\name{cro}
\alias{cro}
\alias{cro_cpct}
\alias{cro_rpct}
\alias{cro_tpct}
\alias{cro_cpct_responses}
\title{Cross tabulation with support of labels, weights and multiple response variables.}
\usage{
cro(cell_vars, col_vars = total(), weight = NULL, subgroup = NULL,
  total_label = "#Total", total_statistic = "unweighted",
  total_row_position = c("below", "above", "none"))

cro_cpct(cell_vars, col_vars = total(), weight = NULL, subgroup = NULL,
  total_label = "#Total", total_statistic = "unweighted",
  total_row_position = c("below", "above", "none"))

cro_rpct(cell_vars, col_vars = total(), weight = NULL, subgroup = NULL,
  total_label = "#Total", total_statistic = "unweighted",
  total_row_position = c("below", "above", "none"))

cro_tpct(cell_vars, col_vars = total(), weight = NULL, subgroup = NULL,
  total_label = "#Total", total_statistic = "unweighted",
  total_row_position = c("below", "above", "none"))

cro_cpct_responses(cell_vars, col_vars = total(), weight = NULL,
  subgroup = NULL, total_label = "#Total", total_statistic = "unweighted",
  total_row_position = c("below", "above", "none"))
}
\arguments{
\item{cell_vars}{vector. By now multiple-response predictor is not supported.}

\item{col_vars}{vector. By now multiple-response predictor is not supported.}

\item{weight}{numeric vector. Optional case weights. NA's and negative weights
treated as zero weights.}

\item{total_row_position}{Position ot total row in the resulting table. Can
be one of "below", "above", "none".}

\item{x}{vector/data.frame. data.frames are considered as multiple response
variables.}

\item{weighted_total}{By default it is "unweighted". Can be "unweighted" or/and "weighted". You can show weighted total/unweighted total or both.}

\item{total_row_title}{By default "#Total". You can provide two titles - for weighted and unqeighted totals.}

\item{fun}{custom summary function. It should always return
scalar/vector/matrix of the same size.}

\item{...}{further arguments for \code{fun}}
}
\value{
object of class 'etable'. Basically it's a data.frame but class
  is needed for custom methods.
}
\description{
\itemize{
\item{\code{cro}}{ returns data.frame with counts (possibly weighted) with column and
row totals.}
\item{\code{cro_pct}, \code{cro_cpct}, \code{cro_rpct}}{ return data.frame with 
table/column/row percent with column and row totals. There are always 
weighted counts instead of margin with 100\%. Empty labels/factor levels are 
removed from results of these functions. Base for multiple response (x is 
data.frame) percent is number of valid cases (not sum of responses) so sum of
percent may be greater than 100. Case is considered as valid if it has at
least one non-NA value.}
\item{\code{cro_mean}, \code{cro_sum}, \code{cro_median}}{ return data.frame with 
mean/sum/median. Empty labels/factor levels are 
removed from results of these functions. NA's are always omitted.}
\item{\code{cro_fun}, \code{cro_fun_df}}{ return data.frame with custom 
summary statistics defined by 'fun' argument. Empty labels/factor levels in 
predictor are removed from results of these functions. NA's treatment depends
on your 'fun' behavior. To use weight you should have 'weight' argument in 
'fun' and some logic for its proccessing inside.\code{cro_fun} applies 'fun'
on each column in 'x' separately, \code{cro_fun_df} gives to 'fun' x as a
whole data.frame. So \code{cro_fun(iris[, -5], iris$Species, fun = mean)}
gives the same result as \code{cro_fun_df(iris[, -5], iris$Species, fun =
colMeans)}. For \code{cro_fun_df} names of 'x' will converted to labels if
they are available before 'fun' is applied. You should take care to return
from 'fun' rectangular object with appropriate row/column names - they will
be used in final result as labels.}
}
}
\examples{
data(mtcars)
mtcars = apply_labels(mtcars,
                      mpg = "Miles/(US) gallon|Mean",
                      cyl = "Number of cylinders",
                      disp = "Displacement (cu.in.)|Mean",
                      hp = "Gross horsepower|Mean",
                      drat = "Rear axle ratio",
                      wt = "Weight (lb/1000)",
                      qsec = "1/4 mile time|Mean",
                      vs = "Engine",
                      vs = c("V-engine" = 0,
                             "Straight engine" = 1),
                      am = "Transmission",
                      am = c("Automatic" = 0,
                             "Manual"=1),
                      gear = "Number of forward gears",
                      carb = "Number of carburetors"
)

calculate(mtcars, cro(am, vs))

# multiple banners
calculate(mtcars, cro(cyl, list("#Total", vs, am, carb)))

# column percent
calculate(mtcars, cro_cpct(cyl, list("#Total", vs, am, carb)))

# nested banner
calculate(mtcars, cro(cyl, list("#Total", vs \%nest\% am, carb)))

# stacked variables
calculate(mtcars, cro(list(cyl, carb), list("#Total", vs \%nest\% am)))

# nested variables
calculate(mtcars, cro(am \%nest\% cyl, list("#Total", vs, carb)))

# multiple-choice variable
# brands - multiple response question
# Which brands do you use during last three months? 
set.seed(123)
brands = data.frame(t(replicate(20,sample(c(1:5,NA),4,replace = FALSE))))
# score - evaluation of tested product
score = sample(-1:1,20,replace = TRUE)
var_lab(brands) = "Used brands"
val_lab(brands) = make_labels("
                              1 Brand A
                              2 Brand B
                              3 Brand C
                              4 Brand D
                              5 Brand E
                              ")

var_lab(score) = "Evaluation of tested brand"
val_lab(score) = make_labels("
                             -1 Dislike it
                             0 So-so
                             1 Like it    
                             ")

cro(mrset(brands), list("#Total", score))
cro_cpct(mrset(brands), list("#Total", score))

# 'cro_mean'

data(iris)
cro_mean(iris[, -5], iris$Species)

# 'cro_fun'

data(mtcars)
mtcars = modify(mtcars,{
    var_lab(vs) = "Engine"
    val_lab(vs) = c("V-engine" = 0, 
                    "Straight engine" = 1) 
    var_lab(hp) = "Gross horsepower"
    var_lab(mpg) = "Miles/(US) gallon"
})

# Label for 'disp' forgotten intentionally
with(mtcars, cro_fun(data.frame(hp, mpg, disp), vs, summary))

# or, the same with transposed summary
with(mtcars, cro_fun(data.frame(hp, mpg, disp), vs, function(x) t(summary(x))))

# very artificial example
a = c(1,1,1, 1, 1)
b = c(0, 1, 2, 2, NA)
weight = c(0, 0, 1, 1, 1)
cro_fun(b, a, weight = weight, 
     fun = function(x, weight, na.rm){
                 weighted.mean(x, w = weight, na.rm = na.rm)
             }, 
     na.rm = TRUE)


# comparison 'cro_fun' and 'cro_fun_df'

data(iris)
cro_fun(iris[, -5], iris$Species, fun = mean)
# same result
cro_fun_df(iris[, -5], iris$Species, fun = colMeans)  

# usage for 'cro_fun_df' which is not possible for 'cro_fun'
# calculate correlations of variables with Sepal.Length inside each group
cro_fun_df(iris[,-5], iris$Species, fun = function(x) cor(x)[,1])

# or, pairwise correlations inside groups
cro_fun_df(iris[,-5], iris$Species, fun = cor)
}
