% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_datatable.R
\name{datatable}
\alias{datatable}
\alias{datatable.etable}
\title{Create an HTML table widget for usage with Shiny}
\usage{
datatable(data, ...)

\method{datatable}{etable}(data, repeat_row_labels = FALSE,
  show_row_numbers = FALSE, digits = getOption("expss.digits"), ...)
}
\arguments{
\item{data}{a data object (result of \link{fre}/\link{cro} and etc)}

\item{...}{further parameters for \link[DT]{datatable}}

\item{repeat_row_labels}{logical Should we repeat duplicated row labels in
the every row? Default is FALSE.}

\item{show_row_numbers}{logical Default is FALSE.}

\item{digits}{integer By default, all numeric columns are rounded to one digit after
decimal separator. Also you can set this argument by option 'expss.digits'
- for example, \code{option(expss.digits = 2)}. If it is NA than all
numeric columns remain unrounded.}
}
\value{
Object of class \link[DT]{datatable}
}
\description{
This is method for rendering results of \link{fre}/\link{cro} in Shiny.
For detailed description of function and its arguments see \link[DT]{datatable}.
}
\examples{
\dontrun{ 
data(mtcars)
mtcars = apply_labels(mtcars,
                      mpg = "Miles/(US) gallon|Mean",
                      cyl = "Number of cylinders",
                      disp = "Displacement (cu.in.)|Mean",
                      hp = "Gross horsepower|Mean",
                      drat = "Rear axle ratio",
                      wt = "Weight (lb/1000)",
                      qsec = "1/4 mile time|Mean",
                      vs = "Engine",
                      vs = c("V-engine" = 0,
                             "Straight engine" = 1),
                      am = "Transmission",
                      am = c("Automatic" = 0,
                             "Manual"=1),
                      gear = "Number of forward gears",
                      carb = "Number of carburetors"
)

mtcars_table = calculate(mtcars,
                         cro_mean(list(mpg, hp), am \%nest\% vs) \%add_rows\%
                             cro_cpct(am, am \%nest\% vs)  
                         
)
library(shiny)
shinyApp(
    ui = fluidPage(fluidRow(column(12, DT::dataTableOutput('tbl')))),
    server = function(input, output) {
        output$tbl = DT::renderDataTable(
            datatable(mtcars_table)
        )
    }
)
}
}
\seealso{
\link[htmlTable]{htmlTable} for knitting
}
