% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dichotomy.R
\name{dichotomy}
\alias{dichotomy}
\title{Convert variable to matrix of dummy variables (dichotomy).}
\usage{
dichotomy(x, keep_unused = FALSE, use_na = TRUE, keep = NULL)
}
\arguments{
\item{x}{vector/factor/matrix/data.frame.}

\item{keep_unused}{Logical. Should we create columns for unused labels/factor levels.}

\item{use_na}{Logical. Should we use NA for rows with all NA or use 0.}

\item{keep}{Numeric/character. Labels/levels/values that should be keeped.
By default all used values will be keeped.}
}
\value{
matrix with 0,1 which column names are value labels. If label doesn't exist for
particular value then this value will be used as column name.
}
\description{
\code{dichotomy} returns matrix with 0,1 and possibly NA.
}
\details{
.
}
\examples{
test_ds = data.frame(total = 1, s2b = sample(2:3,100,replace = TRUE))
age_group = test_ds$s2b
var_lab(test_ds$s2b) = "Age group"
age_group = set_var_lab(age_group,"Age group")
identical(age_group,test_ds$s2b) # should be TRUE

identical(var_lab(age_group),attr(age_group,"label")) # should be TRUE
}
\seealso{
\code{\link{category}} for reverse conversion.
}

