% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_repeat.R
\name{do_repeat}
\alias{.do_repeat}
\alias{do_repeat}
\title{Repeats the same transformations on a specified set of variables/values}
\usage{
do_repeat(data, ...)

.do_repeat(...)
}
\arguments{
\item{data}{data.frame}

\item{...}{}
}
\value{
transformed data.frame \code{data}
}
\description{
Repeats the same transformations on a specified set of variables/values
}
\examples{
1
}

