% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_empty.R
\name{drop_empty_rows}
\alias{drop_c}
\alias{drop_empty_columns}
\alias{drop_empty_rows}
\alias{drop_r}
\alias{drop_rc}
\title{Drop empty (with all NA's) rows/columns from data.frame/table}
\usage{
drop_empty_rows(x, excluded_columns = 1, excluded_rows = grep("^#", x[[1]],
  perl = TRUE))

drop_empty_columns(x, excluded_columns = 1, excluded_rows = grep("^#",
  x[[1]], perl = TRUE))

drop_r(x, excluded_columns = 1, excluded_rows = grep("^#", x[[1]], perl =
  TRUE))

drop_c(x, excluded_columns = 1, excluded_rows = grep("^#", x[[1]], perl =
  TRUE))

drop_rc(x)
}
\arguments{
\item{x}{data.frame/column}

\item{excluded_columns}{logical/numeric/characters columns which won't be
dropped and in which NAs won't be counted. By default, it is first column -
column with labels in table.}

\item{excluded_rows}{logical/numeric rows which won't be dropped and in which
NAs won't be counted. By default, it is rows which have values with "#" at
the beginning  in the first column. it is total rows in tables.}
}
\value{
data.frame with removed rows/columns
}
\description{
By default tables produced by functions \link{table_cases}, 
\link{table_summary} and \link{table_summary_df} are created with all 
possible value labels. If values of this labels are absent in variable there 
are NA's in rows and columns. 
\code{drop_empty_rows}/\code{drop_empty_columns} are intended to remove 
these empty rows/columns. \code{drop_r} and \code{drop_c} are the same
functions with shorter names. \code{drop_rc} drops rows and columns
simultaneously.
}
\examples{
data(mtcars)
mtcars = apply_labels(mtcars,
            vs = "Engine",
            vs = num_lab("
                      0 V-engine 
                      1 Straight engine
                      9 Other
                      "),
            am = "Transmission",
            am = num_lab("
                     0 Automatic 
                     1 Manual
                     9 Other
                     ")
         )
with_empty_rows = calculate(mtcars, table_cases(am, vs))

drop_empty_rows(with_empty_rows)
drop_empty_columns(with_empty_rows)
drop_rc(with_empty_rows)
                        
}

