% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fre.R
\name{fre}
\alias{cro}
\alias{fre}
\title{Title}
\usage{
fre(x, weight = NULL)

cro(x, predictor, weight = NULL)
}
\arguments{
\item{x}{}

\item{weight}{}

\item{predictor}{}
}
\value{
data.frame
}
\description{
Title
}
\examples{
a = a + 1
}

