% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{get_var_range}
\alias{\%to\%}
\alias{\%to_list\%}
\alias{get_var_range}
\alias{get_var_range_list}
\alias{get_vars}
\alias{get_vars_list}
\title{Get range of variables/variables by pattern}
\usage{
get_var_range(start, end)

get_var_range_list(start, end)

e1 \%to\% e2

e1 \%to_list\% e2

get_vars_list(pattern)

get_vars(pattern)
}
\arguments{
\item{start}{character Name of start variable (e. g. a_1)}

\item{end}{character Name of start variable (e. g. a_5)}

\item{e1}{unquoted name of start variable (e. g. a_1)}

\item{e2}{unquoted name of start variable (e. g. a_5)}

\item{pattern}{character pattern of variable(-s) name}
}
\value{
data.frame/list with variables
}
\description{
Get range of variables/variables by pattern
}
\examples{
a = 1
}

