% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{htmlTable.etable}
\alias{htmlTable.etable}
\title{Outputting HTML tables in RStudio viewer/R Notebooks}
\usage{
\method{htmlTable}{etable}(x, digits = getOption("expss.digits"), ...)
}
\arguments{
\item{x}{a data object (result of \link{fre}/\link{cro} and etc)}

\item{digits}{integer By default, all numeric columns are rounded to one digit after
decimal separator. Also you can set this argument by option 'expss.digits'
- for example, \code{option(expss.digits = 2)}. If it is NA than all
numeric columns remain unrounded.}

\item{...}{further parameters for \link[htmlTable]{htmlTable}.}
}
\value{
Returns a string of class htmlTable
}
\description{
This is method for rendering results of \link{fre}/\link{cro} in
Shiny/RMarkdown and etc. For detailed description of function and its
arguments see \link[htmlTable]{htmlTable}. You may be interested in
\code{options(expss.print_table = "viewer")} for automatical rendering tables
in the RStudio viewer or  \code{options(expss.print_table = "rnotebook")} for
rendering in the R notebooks.
}
\examples{
a = 1
}
