% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrset.R
\name{mrset}
\alias{mdset}
\alias{mrset}
\title{Create multiple response set/multiple dichotomy set from variables}
\usage{
mrset(...)

mdset(...)
}
\arguments{
\item{...}{variables}
}
\value{
data.frame of class \code{category}/\code{dichotomy}
}
\description{
This functions are intended for usage with tables - \link{table_cases}, 
\link{table_summary}. Result of \code{mrset} is considered as 
muliple-response set with category encoding and result of \code{mdset} is 
considered as multiple response set with dichotomy encoding e. g. with 0 or 1
in the each column. Each column in the \code{dichotomy} is indicator of 
absense or presense of particular feature. Both functions don't convert its 
arguments to anything - it is supposed that arguments already have 
appropriate encoding. For conversation see \link{dichotomy} or
\link{category}.
}
\examples{

data(product_test)
# a1 = with(product_test, mrset(a2_1 \%to\% a2_99))

}

