% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expss.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%<>\%}
\alias{\%$\%}
\alias{data.table}
\alias{as.data.table}
\alias{is.data.table}
\alias{setDF}
\alias{setDT}
\alias{setkey}
\alias{setkeyv}
\alias{fread}
\alias{fwrite}
\alias{shift}
\alias{first}
\alias{last}
\alias{setorder}
\alias{setcolorder}
\alias{between}
\alias{chgroup}
\alias{chmatch}
\alias{chorder}
\alias{CJ}
\alias{fintersect}
\alias{foverlaps}
\alias{frank}
\alias{frankv}
\alias{fsetdiff}
\alias{fsetequal}
\alias{fsort}
\alias{funion}
\alias{getDTthreads}
\alias{getNumericRounding}
\alias{haskey}
\alias{indices}
\alias{inrange}
\alias{key}
\alias{like}
\alias{rbindlist}
\alias{rleid}
\alias{rleidv}
\alias{rowid}
\alias{rowidv}
\alias{set}
\alias{set2key}
\alias{set2keyv}
\alias{setattr}
\alias{setDTthreads}
\alias{setindex}
\alias{setindexv}
\alias{setnames}
\alias{setNumericRounding}
\alias{setorderv}
\alias{SJ}
\alias{transpose}
\alias{tstrsplit}
\alias{uniqueN}
\alias{\%between\%}
\alias{\%chin\%}
\alias{\%inrange\%}
\alias{\%like\%}
\alias{htmlTable}
\alias{htmlTableWidget}
\alias{htmlTableWidgetOutput}
\alias{interactiveTable}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{data.table}}, \code{\link[data.table]{as.data.table}}, \code{\link[data.table]{is.data.table}}, \code{\link[data.table]{setDF}}, \code{\link[data.table]{setDT}}, \code{\link[data.table]{setkey}}, \code{\link[data.table]{setkeyv}}, \code{\link[data.table]{fread}}, \code{\link[data.table]{fwrite}}, \code{\link[data.table]{shift}}, \code{\link[data.table]{first}}, \code{\link[data.table]{last}}, \code{\link[data.table]{setorder}}, \code{\link[data.table]{setcolorder}}, \code{\link[data.table]{between}}, \code{\link[data.table]{chgroup}}, \code{\link[data.table]{chmatch}}, \code{\link[data.table]{chorder}}, \code{\link[data.table]{CJ}}, \code{\link[data.table]{fintersect}}, \code{\link[data.table]{foverlaps}}, \code{\link[data.table]{frank}}, \code{\link[data.table]{frankv}}, \code{\link[data.table]{fsetdiff}}, \code{\link[data.table]{fsetequal}}, \code{\link[data.table]{fsort}}, \code{\link[data.table]{funion}}, \code{\link[data.table]{getDTthreads}}, \code{\link[data.table]{getNumericRounding}}, \code{\link[data.table]{haskey}}, \code{\link[data.table]{indices}}, \code{\link[data.table]{inrange}}, \code{\link[data.table]{key}}, \code{\link[data.table]{like}}, \code{\link[data.table]{rbindlist}}, \code{\link[data.table]{rleid}}, \code{\link[data.table]{rleidv}}, \code{\link[data.table]{rowid}}, \code{\link[data.table]{rowidv}}, \code{\link[data.table]{set}}, \code{\link[data.table]{set2key}}, \code{\link[data.table]{set2keyv}}, \code{\link[data.table]{setattr}}, \code{\link[data.table]{setcolorder}}, \code{\link[data.table]{setDTthreads}}, \code{\link[data.table]{setindex}}, \code{\link[data.table]{setindexv}}, \code{\link[data.table]{setkey}}, \code{\link[data.table]{setkeyv}}, \code{\link[data.table]{setnames}}, \code{\link[data.table]{setNumericRounding}}, \code{\link[data.table]{setorder}}, \code{\link[data.table]{setorderv}}, \code{\link[data.table]{SJ}}, \code{\link[data.table]{transpose}}, \code{\link[data.table]{tstrsplit}}, \code{\link[data.table]{uniqueN}}, \code{\link[data.table]{\%between\%}}, \code{\link[data.table]{\%chin\%}}, \code{\link[data.table]{\%inrange\%}}, \code{\link[data.table]{\%like\%}}}

  \item{htmlTable}{\code{\link[htmlTable]{htmlTable}}, \code{\link[htmlTable]{htmlTableWidget}}, \code{\link[htmlTable]{htmlTableWidgetOutput}}, \code{\link[htmlTable]{interactiveTable}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%<>\%}}, \code{\link[magrittr]{\%$\%}}}
}}

