% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expss.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%<>\%}
\alias{\%$\%}
\alias{data.table}
\alias{as.data.table}
\alias{is.data.table}
\alias{setDF}
\alias{setDT}
\alias{setkey}
\alias{setkeyv}
\alias{fread}
\alias{fwrite}
\alias{shift}
\alias{first}
\alias{last}
\alias{setorder}
\alias{setcolorder}
\alias{between}
\alias{chgroup}
\alias{chmatch}
\alias{chorder}
\alias{CJ}
\alias{fintersect}
\alias{foverlaps}
\alias{frank}
\alias{frankv}
\alias{fsetdiff}
\alias{fsetequal}
\alias{fsort}
\alias{funion}
\alias{getDTthreads}
\alias{haskey}
\alias{indices}
\alias{inrange}
\alias{key}
\alias{rbindlist}
\alias{rleid}
\alias{rleidv}
\alias{rowid}
\alias{rowidv}
\alias{set}
\alias{setattr}
\alias{setDTthreads}
\alias{setindex}
\alias{setindexv}
\alias{setnames}
\alias{setorderv}
\alias{SJ}
\alias{transpose}
\alias{tstrsplit}
\alias{uniqueN}
\alias{htmlTable}
\alias{htmlTableWidget}
\alias{htmlTableWidgetOutput}
\alias{interactiveTable}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table:J]{CJ}}, \code{\link[data.table:J]{SJ}}, \code{\link[data.table]{as.data.table}}, \code{\link[data.table]{between}}, \code{\link[data.table:chmatch]{chgroup}}, \code{\link[data.table]{chmatch}}, \code{\link[data.table:chmatch]{chorder}}, \code{\link[data.table]{data.table}}, \code{\link[data.table:setops]{fintersect}}, \code{\link[data.table:last]{first}}, \code{\link[data.table]{foverlaps}}, \code{\link[data.table]{frank}}, \code{\link[data.table:frank]{frankv}}, \code{\link[data.table]{fread}}, \code{\link[data.table:setops]{fsetdiff}}, \code{\link[data.table:setops]{fsetequal}}, \code{\link[data.table]{fsort}}, \code{\link[data.table:setops]{funion}}, \code{\link[data.table]{fwrite}}, \code{\link[data.table:openmp-utils]{getDTthreads}}, \code{\link[data.table:setkey]{haskey}}, \code{\link[data.table:setkey]{indices}}, \code{\link[data.table:between]{inrange}}, \code{\link[data.table:as.data.table]{is.data.table}}, \code{\link[data.table:setkey]{key}}, \code{\link[data.table]{last}}, \code{\link[data.table]{rbindlist}}, \code{\link[data.table]{rleid}}, \code{\link[data.table:rleid]{rleidv}}, \code{\link[data.table]{rowid}}, \code{\link[data.table:rowid]{rowidv}}, \code{\link[data.table:assign]{set}}, \code{\link[data.table]{setDF}}, \code{\link[data.table]{setDT}}, \code{\link[data.table:openmp-utils]{setDTthreads}}, \code{\link[data.table]{setattr}}, \code{\link[data.table]{setcolorder}}, \code{\link[data.table]{setcolorder}}, \code{\link[data.table:setkey]{setindex}}, \code{\link[data.table:setkey]{setindexv}}, \code{\link[data.table]{setkey}}, \code{\link[data.table]{setkey}}, \code{\link[data.table:setkey]{setkeyv}}, \code{\link[data.table:setkey]{setkeyv}}, \code{\link[data.table:setattr]{setnames}}, \code{\link[data.table]{setorder}}, \code{\link[data.table]{setorder}}, \code{\link[data.table:setorder]{setorderv}}, \code{\link[data.table]{shift}}, \code{\link[data.table]{transpose}}, \code{\link[data.table]{tstrsplit}}, \code{\link[data.table:duplicated]{uniqueN}}}

  \item{htmlTable}{\code{\link[htmlTable]{htmlTable}}, \code{\link[htmlTable]{htmlTableWidget}}, \code{\link[htmlTable:htmlTableWidget-shiny]{htmlTableWidgetOutput}}, \code{\link[htmlTable]{interactiveTable}}}

  \item{magrittr}{\code{\link[magrittr:exposition]{\%$\%}}, \code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}}
}}

