% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet.R
\name{sheet}
\alias{sheet}
\alias{as.sheet}
\title{Make data.frame without conversion to factors and without fixing names}
\usage{
sheet(...)

as.sheet(x, ...)
}
\arguments{
\item{...}{objects, possibly named}

\item{x}{object to be coerced to data.frame}
}
\value{
data.frame/list
}
\description{
\code{sheet} and \code{as.sheet} are shortcuts to \code{data.frame} and 
\code{as.data.frame} with stringsAsFactors = FALSE, check.names = FALSE.
}
\examples{

# see the difference
df1 = data.frame(a = letters[1:3], "This is my long name" = 1:3)
df2 = sheet(a = letters[1:3], "This is my long name" = 1:3)

str(df1)
str(df2)


}
\seealso{
\link[base]{data.frame}, \link[base]{as.data.frame}
}
