% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R, R/significance_means.R
\name{significance_cpct}
\alias{significance_cpct}
\alias{add_sig_labels}
\alias{significance_means}
\title{Mark significant differences between columns of the table}
\usage{
significance_cpct(x, sig_level = 0.05, min_base = 2,
  compare_type = "subtable", bonferroni = FALSE, sig_labels = LETTERS,
  sig_labels_previous_column = c("v", "^"), sig_labels_first_column = c("-",
  "+"), na_as_zero = FALSE, total_marker = "#", total_row = 1)

add_sig_labels(tbl, sig_labels = LETTERS)

significance_means(x, sig_level = 0.05, min_base = 2,
  compare_type = "subtable", bonferroni = FALSE, sig_labels = LETTERS,
  sig_labels_previous_column = c("v", "^"), sig_labels_first_column = c("-",
  "+"), stat_order = c("mean", "sd", "n"), var_equal = FALSE)
}
\arguments{
\item{x}{table with proportions and bases - result of \link{cro_cpct} for \code{significance_cpct}}

\item{sig_level}{significance level - by default it equals to \code{0.05}}

\item{min_base}{integer number. Significance test will be conducted if both
columns have bases greater than \code{min_base}. By default it equals to \code{2}.}

\item{compare_type}{Type of compare between columns. By default it is 
\code{subtable} - comparisons will be conducted between columns of each 
subtable. Other possible values are: \code{first_column},
\code{first_column_adjusted} and \code{previous_column}}

\item{bonferroni}{logical. \code{FALSE} by default. Should we use Bonferrony
adjustment for multiple comparisons?}

\item{sig_labels}{character vector. Labels for marking differences between columns of subtable.}

\item{sig_labels_previous_column}{character vector with two elements. Labels
for marking difference with previous column. First mark means 'lower' (by
default it is \code{v}) and the second means greater (\code{^}).}

\item{sig_labels_first_column}{character vector with two elements. Labels
for marking difference with first column of the table. First mark means 'lower' (by
default it is \code{-}) and the second means 'greater' (\code{+}).}

\item{na_as_zero}{logical. \code{FALSE} by default. Should we treat \code{NA}'s as zero cases?}

\item{total_marker}{character. Mark of total rows in table.}

\item{total_row}{integer number. In case of several totals per subtable it is
number of total row for significance calculation.}
}
\value{
Object of class \code{etable} with marks of significant differences between columns.
}
\description{
Mark significant differences between columns of the table
}
\examples{
1
}
