% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R, R/significance_means.R
\name{significance_cpct}
\alias{significance_cpct}
\alias{add_sig_labels}
\alias{compare_proportions}
\alias{compare_means}
\alias{significance_means}
\title{Title}
\usage{
significance_cpct(x, sig_level = 0.05, min_base = 2,
  compare_type = "subtable", bonferroni = FALSE, sig_labels = LETTERS,
  sig_labels_previous_column = c("v", "^"), sig_labels_first_column = c("-",
  "+"), na_as_zero = FALSE, total_marker = "#", total_row = 1)

add_sig_labels(tbl, sig_labels = LETTERS)

compare_proportions(prop1, prop2, base1, base2, common_base = 0)

compare_means(mean1, mean2, sd1, sd2, base1, base2, common_base = 0,
  var_equal = FALSE)

significance_means(x, sig_level = 0.05, min_base = 2,
  compare_type = "subtable", bonferroni = FALSE, sig_labels = LETTERS,
  sig_labels_previous_column = c("v", "^"), sig_labels_first_column = c("-",
  "+"), stat_order = c("mean", "sd", "n"), var_equal = FALSE)
}
\arguments{
\item{x}{gddfg}

\item{sig_level}{dfgdfg}

\item{min_base}{dfgfg}

\item{compare_type}{dfg}

\item{sig_labels}{dfg}

\item{sig_labels_previous_column}{dfgdfg}

\item{sig_labels_first_column}{dfgdfgfd}
}
\value{
fsdgsdg
}
\description{
Title
}
\examples{
1
}
