% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{significance_cpct}
\alias{significance_cpct}
\title{Title}
\usage{
significance_cpct(x, sig_level = 0.05, min_base = 2,
  compare_type = c("subtable", "subtable_bonferrony", "first_column",
  "first_column_adjusted", "previous_column"), sig_labels = LETTERS,
  sig_labels_previous_column = c("-", "+"), sig_labels_first_column = c("-",
  "+"), na_as_zero = FALSE, total_marker = "#", total_row = 1,
  only_labels = FALSE, digits = getOption("expss.digits"))
}
\arguments{
\item{x}{gddfg}

\item{sig_level}{dfgdfg}

\item{min_base}{dfgfg}

\item{compare_type}{dfg}

\item{sig_labels}{dfg}

\item{sig_labels_previous_column}{dfgdfg}

\item{sig_labels_first_column}{dfgdfgfd}
}
\value{
fsdgsdg
}
\description{
Title
}
\examples{
1
}
