% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_summary_df.R
\name{table_summary_df}
\alias{table_pearson}
\alias{table_spearman}
\alias{table_summary_df}
\title{Title}
\usage{
table_summary_df(summary_vars, col_vars, fun, weight = NULL,
  row_vars = NULL, row_labels = c("row_vars", "row_vars_values"),
  col_labels = c("col_vars", "col_vars_values"), hide = NULL,
  use_result_row_order = TRUE)

table_pearson(summary_vars, col_vars, weight = NULL, row_vars = NULL)

table_spearman(summary_vars, col_vars, weight = NULL, row_vars = NULL)
}
\arguments{
\item{summary_vars}{dsd}

\item{col_vars}{dssd}

\item{fun}{dsd}

\item{weight}{sdsd}

\item{row_vars}{dssd}

\item{row_labels}{sdsd}

\item{col_labels}{dssd}

\item{hide}{dssd}

\item{use_result_row_order}{sdsd}
}
\value{
sdsd
}
\description{
Title
}
\examples{
1
}

