% Generated by roxygen2 (4.0.2): do not edit by hand
\name{val_lab}
\alias{add_val_lab}
\alias{add_val_lab<-}
\alias{unvl}
\alias{val_lab}
\alias{val_lab<-}
\alias{with_val_lab}
\title{Set or get value labels}
\usage{
val_lab(x)
}
\arguments{
\item{x}{Variable(s). Vector/data.frame/list.}

\item{value}{Named vector. Names of vector values are labels for the
appropriate values of variable x.}
}
\value{
\code{val_lab} return value labels (named vector). If labels doesn't
  exist it return NULL . \code{val_lab<-} and \code{with_val_lab} return
  variable (vector x) of class "with_labels" with attribute "value_labels"
  which contains value labels.
}
\description{
\code{val_lab} returns value labels or NULL if labels doesn't
exist.
\code{val_lab<-} set value labels.
\code{with_val_lab} returns variable with value labels.
\code{add_val_lab<-} add value labels to already existing value labels.
\code{add_val_lab} return variable with added value labels to already existing value labels.
\code{unvl} drops value labels.
}
\details{
Value labels are stored in attribute "value_labels"
  (\code{attr(x,"value_labels")}). Duplicated values are not allowed. If
  argument \code{x} is data.frame or list then labels applied to all elements
  of data.frame/list. We set variable class to "with_labels" for preserving
  from droping this attribute during some operations (such as \code{c}).
  There are special methods of subsetting and concatenation for this class.
  To drop value labels use \code{val_lab(var) <- NULL} or \code{unvl(var)}.
}
\examples{
data(mtcars)
}

