% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/labels.R
\name{var_lab}
\alias{set_var_lab}
\alias{unvr}
\alias{var_lab}
\alias{var_lab<-}
\title{Set or get variable label}
\usage{
var_lab(x)

var_lab(x) <- value

set_var_lab(x, value)

unvr(x)
}
\arguments{
\item{x}{Variable. In the most cases it is numeric vector.}

\item{value}{A character scalar - label for the variable x.}
}
\value{
\code{var_lab} return variable label. If label doesn't exist it return
  NULL . \code{var_lab<-} and \code{set_var_lab} return variable (vector x)
  of class "with_labels" with attribute "label" which equals submitted value.
}
\description{
\code{var_lab} returns variable label or NULL if label doesn't
exist.
\code{var_lab<-} set variable label.
\code{set_var_lab} returns variable with label.
\code{unvr} drops variable label.
}
\details{
Variable label is stored in attribute "label" (\code{attr(x,"label")}). For
  preserving from dropping this attribute during some operations (such as \code{c})
  variable class is set to "with_labels". There are special methods of
  subsetting and concatenation for this class. To drop variable label use
  \code{var_lab(var) <- NULL} or \code{unvr(var)}.
}
\examples{
test_ds = data.frame(total = 1, s2b = sample(2:3,100,replace = TRUE))
age_group = test_ds$s2b
var_lab(test_ds$s2b) = "Age group"
age_group = set_var_lab(age_group,"Age group")
identical(age_group,test_ds$s2b) # should be TRUE

identical(var_lab(age_group),attr(age_group,"label")) # should be TRUE
}

