% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_labels.R
\name{write_labels}
\alias{read_labelled_csv}
\alias{write_labelled_csv}
\alias{write_labelled_spss}
\alias{write_labels}
\alias{write_labels_spss}
\title{Write data labels to file in R code or in SPSS syntax.}
\usage{
write_labels(x, filename, fileEncoding = "")

read_labelled_csv(filename, fileEncoding = "", ...)

write_labelled_csv(x, filename, fileEncoding = "", ...)

write_labelled_spss(x, filename, fileEncoding = "", ...)

write_labels_spss(x, filename)
}
\arguments{
\item{x}{data.frame to be written/data.frame whose labels to be written}

\item{filename}{the name of the file which the data are to be read from/write to.}

\item{fileEncoding}{character string: if non-empty declares the encoding to
be used on a file (not a connection) so the character data can be
re-encoded as they are written. See \link[base]{file}.}

\item{...}{additional arguments for \link[utils]{read.table}/\link[utils]{write.table}}
}
\value{
Functions for writing invisibly return NULL. Functions for reading return labelled data.frame.
}
\description{
Write data labels to file in R code or in SPSS syntax.
}
\examples{
a = 2
}

